% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{abisko}
\alias{abisko}
\title{Abisko rainfall}
\format{
a data frame with 15132 rows and two variables
}
\source{
Abisko Scientific Research Station
}
\arguments{
\item{date}{\code{Date} of the measurement}

\item{precip}{rainfall amount (in mm)}
}
\description{
Daily non-zero rainfall measurements in Abisko (Sweden) from January 1913 until December 2014.
}
\references{
A. Kiriliouk, H. Rootzen, J. Segers and J.L. Wadsworth (2019), \emph{Peaks over thresholds modeling with multivariate generalized Pareto distributions},  Technometrics, \bold{61}(1), 123--135, <doi:10.1080/00401706.2018.1462738>
}
