% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specdens.R
\name{angmeasdir}
\alias{angmeasdir}
\title{Dirichlet mixture model for the spectral density}
\usage{
angmeasdir(x, th, Rnorm = c("l1", "l2", "linf"), Anorm = c("l1", "l2",
  "linf", "arctan"), marg = c("Frechet", "Pareto"), wgt = c("Empirical",
  "Euclidean"), region = c("sum", "min", "max"), which.method = 1)
}
\description{
This function uses the points returned from a call to \code{angmeas}
}
