% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extcoef.R
\name{BR.extcoeffun}
\alias{BR.extcoeffun}
\title{Extremal coefficient function for the Brown-Resnick process}
\usage{
BR.extcoeffun(h, vario, ...)
}
\arguments{
\item{h}{pairwise distance matrix}

\item{vario}{semivariogram function}

\item{...}{additional parameters passed to \code{vario}}
}
\description{
Extremal coefficient function for the Brown-Resnick process
}
\examples{
extcoef <- BR.extcoeffun(seq(0, 20, by = 0.25), vario = function(x){2*x^0.7})
plot(extcoef, type = 'l', ylim = c(1,2))
}
