% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{build.blocks}
\alias{build.blocks}
\title{Compute block maxima and order them by block}
\usage{
build.blocks(xdat, block = 1L, m = 2L)
}
\arguments{
\item{xdat}{vector of length \code{n}}

\item{block}{integer, size of block over which to compute maxima}

\item{m}{number of columns for further sub-blocking}
}
\value{
a matrix with \eqn{\lfloor n/b \rfloor} observations, ordered by row, with \code{m} columns.
}
\description{
Given a time series of observations in \code{xdat},
compute the maximum of blocks of size \code{block} (\eqn{b}),
and then order them by further blocks of size \code{m},
increasing by row from left to right. If the length of
\code{xdat} is not a multiple of \code{block},
the last observations are discarded without warning.
}
