% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cheeseborowind}
\alias{cheeseborowind}
\title{Cheeseboro wind speed data}
\format{
A data frame with 992 rows and 3 variables:
\describe{
  \item{\code{date}}{date of observation}
  \item{\code{direction}}{angle (in degrees) of the wind}
  \item{\code{gust}}{maximum daily wind speed (in meters per second)}
}
}
\source{
Raw US Climate Archive, \url{https://raws.dri.edu/cgi-bin/rawMAIN.pl?caCCHB}, maintained by the Western Regional Climate Center, Desert Research Institute based in Reno, Nevada
}
\usage{
cheeseborowind
}
\description{
Daily measurements of wind speed during the month of January and including
February 1st, from the Cheeseboro (California) weather station.
between 1996 and 2026.
}
\keyword{datasets}
