% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condextr.R
\name{condex.pseudoll}
\alias{condex.pseudoll}
\title{Conditional extremes pseudo log likelihood}
\usage{
condex.pseudoll(par, xdat, index = 1, thresh = 0.95, constrain = FALSE, ...)
}
\arguments{
\item{par}{vector of parameters, \eqn{\alpha}, \eqn{\beta}}

\item{xdat}{matrix or data frame of observations on the Laplace scale}

\item{index}{vector of indices over which to condition, by default the first component}

\item{thresh}{quantile level (uniform scale)}

\item{constrain}{logical; if \code{TRUE}, apply the constraints from Keef, Papastathopoulos and Tawn (2013)}

\item{...}{additional arguments, currently ignored}

\item{type}{estimating equation, either \code{norm} for normal or \code{skewnorm} for skew-normal distribution}
}
\value{
a numeric value with the pseudo log likelihood, with additional arguments \code{"loc"} and \code{"scale"}
}
\description{
Conditional extremes pseudo log likelihood
}
\keyword{internal}
