% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.diag.R
\name{D.diag}
\alias{D.diag}
\title{Robust selection p-value path}
\usage{
D.diag(xdat, u, B = 99L)
}
\arguments{
\item{xdat}{vector of observations}

\item{u}{vector of candidate thresholds}
}
\description{
Robust selection p-value path
}
