% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wdiag.R
\name{.C1}
\alias{.C1}
\title{Contrast matrix}
\usage{
.C1(k)
}
\arguments{
\item{k}{number of columns (the number of rows is \code{k-1})}
}
\value{
a \code{k-1} x \code{k} contrast matrix
}
\description{
Produces a contrast matrix with (1,-1) elements running down the two diagonals
}
\keyword{internal}
