% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCdiag.R
\name{.exp_info_algebraic}
\alias{.exp_info_algebraic}
\title{Algebraic calculation of the expected information}
\usage{
.exp_info_algebraic(x, w, v, m)
}
\arguments{
\item{x}{parameter vector: (\code{sigma1}, \code{phi_1}, \ldots, \code{phi_m})}

\item{w}{differences between thresholds (\code{w[m]} not used)}

\item{v}{thresholds relative to lowest threshold}

\item{m}{number of thresholds}
}
\description{
Algebraic calculation of the expected information
}
\keyword{internal}
