% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{.gev.postpred}
\alias{.gev.postpred}
\title{Posterior predictive distribution and density for the GEV distribution}
\usage{
.gev.postpred(x, posterior, Nyr = 100, type = c("density", "quantile"))
}
\description{
This function calculates the posterior predictive density at points x
based on a matrix of posterior density parameters
}
