% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{.gpd_grimshaw}
\alias{.gpd_grimshaw}
\title{GP fitting function of Grimshaw (1993)}
\usage{
.gpd_grimshaw(x)
}
\arguments{
\item{x}{sample values}
}
\value{
a list with the maximum likelihood estimates of components \code{a} and \code{k}
}
\description{
Function for estimating parameters \code{k} and \code{a} for a random sample from a GPD.
}
\author{
Scott D. Grimshaw
}
