% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmev.R
\name{.mvasym.check}
\alias{.mvasym.check}
\title{Internal function}
\usage{
.mvasym.check(asy, dep, d, model = c("alog", "aneglog", "maxlin"))
}
\arguments{
\item{asy}{a list of \eqn{2^d-1} asymmetry components, as in Stephenson bvevd functions}

\item{dep}{vector of \eqn{2^d-d-1} values for the dependence parameter}

\item{d}{dimension of the model}

\item{model, }{either \code{alog} for the asymmetric logistic or \code{aneglog}
for the asymmetric negative logistic}
}
\value{
a matrix of asymmetry components, enumerating all possible \eqn{2^d-1} subsets of
the power set
}
\description{
Takes a list of asymmetry parameters with an associated dependence vector and returns
the corresponding asymmetry matrix for the asymmetric logistic and asymmetric negative logistic models
}
\details{
This function is extracted from the evd package and modified
(C) Alec Stephenson
}
\keyword{internal}
