% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rbilogspec}
\alias{.rbilogspec}
\title{Generates from \eqn{Q_i}{Qi}, the spectral measure of the bilogistic model}
\usage{
.rbilogspec(n, alpha)
}
\arguments{
\item{n}{sample size}

\item{alpha}{vector of parameter of dimension \code{d}}
}
\value{
an \code{n} by \code{d} sample from the spectral distribution
}
\description{
Simulation algorithm of Boldi (2009) for the bilogistic model
}
\references{
Boldi (2009). A note on the representation of parametric models
for multivariate extremes. \emph{Extremes} \bold{12}, 211--218.
}
\keyword{internal}
