% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rdirmixspec}
\alias{.rdirmixspec}
\title{Generates from \eqn{Q_i}{Qi}, the spectral measure of the Dirichlet mixture model}
\usage{
.rdirmixspec(n, d, alpha, weight)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension of the 1-sample}

\item{alpha}{a \eqn{d \times n} dimensional vector of positive parameter values for the Dirichlet vector}

\item{weight}{a \code{m} vector of mixture weights, which sum to 1}
}
\value{
an \code{n} by \code{d} sample from the spectral distribution
}
\description{
Simulation algorithm of Dombry et al. (2015)
}
\keyword{internal}
