% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rexstudspec}
\alias{.rexstudspec}
\title{Generates from \eqn{Q_i}{Qi}, the spectral measure of the extremal Student model}
\usage{
.rexstudspec(n, sigma, al)
}
\arguments{
\item{al}{the alpha parameter in Proposition 7. Corresponds to degrees of freedom - 1}

\item{index}{index of the location. An integer in {0, ..., \eqn{d-1}}}

\item{Sigma}{a positive semi-definite covariance matrix with unit variance}
}
\value{
an \code{n} by \code{d} sample from the spectral distribution
}
\description{
Generates from \eqn{Q_i}{Qi}, the spectral measure of the extremal Student model
}
\keyword{internal}
