% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rhrspec}
\alias{.rhrspec}
\title{Generates from \eqn{Q_i}{Qi}, the spectral measure of the Husler-Reiss model}
\usage{
.rhrspec(n, Lambda)
}
\arguments{
\item{Lambda}{an symmetric square matrix of coefficients \eqn{\lambda^2}}
}
\value{
an \code{n} by \code{d} sample from the spectral distribution
}
\description{
Generates from \eqn{Q_i}{Qi}, the spectral measure of the Husler-Reiss model
}
\keyword{internal}
