% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rmevasy}
\alias{.rmevasy}
\title{Random samples from asymmetric logistic distribution}
\usage{
.rmevasy(n, d, par, asym, ncompo, Sigma, model)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension of the multivariate distribution}

\item{par}{a vector of parameters}

\item{asym}{matrix of bool indicating which component belong to the corresponding row logistic model}

\item{ncompo}{number of components for the (negative) logistic in row}

\item{Sigma}{matrix of asymmetry parameters}
}
\value{
a \code{n} by \code{d} matrix containing the sample
}
\description{
Simulation algorithm of Stephenson (2003), using exact-samples from the logistic
}
\references{
Stephenson, A. G. (2003) Simulating multivariate extreme value distributions of logistic type.
\emph{Extremes}, \bold{6}(1), 49--60.

Joe, H. (1990). Families of min-stable multivariate exponential and multivariate
extreme value distributions, \bold{9}, 75--81.
}
\keyword{internal}
