% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rmnorm_arma}
\alias{.rmnorm_arma}
\title{Multivariate Normal distribution sampler (Rcpp version), derived using the eigendecomposition
of the covariance matrix Sigma. The function utilizes the arma random normal generator}
\usage{
.rmnorm_arma(n, Mu, Xmat, eigen = TRUE)
}
\arguments{
\item{n}{sample size}

\item{Mu}{mean vector. Will set the dimension}

\item{Xmat}{covariance matrix, of same dimension as \code{Mu} (and square matrix).
No sanity check is performed to validate that the matrix is symmetric, so use at own risk}
}
\value{
an \code{n} sample from a multivariate Normal distribution
}
\description{
Multivariate Normal distribution sampler (Rcpp version), derived using the eigendecomposition
of the covariance matrix Sigma. The function utilizes the arma random normal generator
}
\keyword{internal}
