% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rneglogspec}
\alias{.rneglogspec}
\title{Generates from \eqn{Q_i}{Qi}, the spectral measure of the negative logistic model}
\usage{
.rneglogspec(n, d, theta)
}
\arguments{
\item{n}{sample size}

\item{theta}{a one-dimensional parameter}
}
\value{
an \code{n} by \code{d} sample from the spectral distribution
}
\description{
Simulation algorithm of Dombry et al. (2016)
}
\references{
Dombry, Engelke and Oesting (2016). Exact simulation of max-stable processes,
\emph{Biometrika}, \bold{103}(2), 303--317.
}
\keyword{internal}
