% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rpairexpspec}
\alias{.rpairexpspec}
\title{Generates from \eqn{Q_i}{Qi}, the spectral measure of the pairwise exponential model}
\usage{
.rpairexpspec(n, d, alpha, beta)
}
\arguments{
\item{n}{sample size}

\item{alpha}{concentration parameter}

\item{beta}{vector of all pairwise component (lexicographic order, by row)}
}
\value{
a matrix of samples from the angular distribution

an \code{n} by \code{d} sample from the spectral distribution
}
\description{
The sample is drawn from a mixture and the algorithm follows from the proof of Theorem 1 in Ballani and Schlather (2011).
}
\references{
Ballani, D. and M. Schlather (2011). A construction principle for multivariate extreme value distributions, \emph{Biometrika}, \bold{98}(3), 633--645.
}
\author{
Leo Belzile
}
\keyword{internal}
