% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rPbilog}
\alias{.rPbilog}
\title{Generate from bilogistic \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is the probability of extremal functions}
\usage{
.rPbilog(d, index, alpha)
}
\arguments{
\item{d}{dimension of the 1-sample}

\item{index}{index of the location. An integer in {0, ..., \eqn{d-1}}}

\item{alpha}{a \eqn{d} dimensional vector of positive parameter values for the Dirichlet vector}
}
\value{
a \code{d}-vector from \eqn{P_x}
}
\description{
Generate from bilogistic \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is the probability of extremal functions
}
\keyword{internal}
