% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rPBrownResnick}
\alias{.rPBrownResnick}
\title{Generate from Brown-Resnick process \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function}
\usage{
.rPBrownResnick(index, Sigma_chol, Sigma)
}
\arguments{
\item{index}{index of the location. An integer in 0, ..., \eqn{d-1}}

\item{Sigma}{a positive definite covariance matrix}
}
\value{
a \code{d}-vector from \eqn{P_x}
}
\description{
Generate from Brown-Resnick process \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function
}
\keyword{internal}
