% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rPdir}
\alias{.rPdir}
\title{Generate from extremal Dirichlet \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal functions from the Dirichlet model of
Coles and Tawn.}
\usage{
.rPdir(d, index, alpha, irv = FALSE)
}
\arguments{
\item{d}{dimension of the 1-sample}

\item{index}{index of the location. An integer in {0, ..., \eqn{d-1}}}

\item{alpha}{a \eqn{d} dimensional vector of positive parameter values for the Dirichlet vector, or
\eqn{d+1} if the last entry is the index of regular variation of the model, a constant in \code{(0, 1]}}

\item{irv}{should the usual model (\code{FALSE}) or the general scaled version (\code{TRUE}) be used}
}
\value{
a \code{d}-vector from \eqn{P_x}
}
\description{
Note: we generate from the Dirichlet rather than the Gamma distribution, since the former is parallelized
}
\keyword{internal}
