% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rPdirmix}
\alias{.rPdirmix}
\title{Generate from extremal Dirichlet \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is the probability of extremal functions from a Dirichlet mixture}
\usage{
.rPdirmix(d, index, alpha, weight)
}
\arguments{
\item{d}{dimension of the 1-sample}

\item{index}{index of the location. An integer in {0, ..., \eqn{d-1}}}

\item{alpha}{a \eqn{d \times n} dimensional vector of positive parameter values for the Dirichlet vector}

\item{weight}{a \code{m} vector of mixture weights, which sum to 1}
}
\value{
a \code{d}-vector from \eqn{P_x}
}
\description{
Generate from extremal Dirichlet \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is the probability of extremal functions from a Dirichlet mixture
}
\keyword{internal}
