% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rPHuslerReiss}
\alias{.rPHuslerReiss}
\title{Generate from extremal Husler-Reiss distribution \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function}
\usage{
.rPHuslerReiss(index, cholesky, Sigma)
}
\arguments{
\item{index}{index of the location. An integer in {0, ..., \eqn{d-1}}}

\item{cholesky}{the Cholesky root of \code{Sigma}}

\item{Sigma}{a covariance matrix formed from the symmetric square matrix of coefficients \eqn{\lambda^2}}
}
\value{
a \code{d}-vector from \eqn{P_x}
}
\description{
Generate from extremal Husler-Reiss distribution \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function
}
\keyword{internal}
