% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rPneglog}
\alias{.rPneglog}
\title{Generate from negative logistic \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function scaled by a Frechet variate}
\usage{
.rPneglog(d, index, theta)
}
\arguments{
\item{d}{dimension of the 1-sample}

\item{index}{index of the location. An integer in {0, ..., \eqn{d-1}}}

\item{theta}{a one-dimensional parameter for the negative logistic model}
}
\value{
a \code{d}-vector from \eqn{P_x}
}
\description{
Generate from negative logistic \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function scaled by a Frechet variate
}
\keyword{internal}
