% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rPsite}
\alias{.rPsite}
\title{Samples from exceedances at site (scaled extremal function definition)}
\usage{
.rPsite(n, j, d, para, model, Sigma, loc)
}
\arguments{
\item{n}{sample size}

\item{d}{dimension of the multivariate distribution}

\item{model}{integer, currently ranging from 1 to 9, corresponding respectively to
(1) \code{log}, (2) \code{neglog}, (3) \code{dirmix}, (4) \code{bilog},
(5) \code{extstud}, (6) \code{br}, (7) \code{ct} and \code{sdir}, (8) \code{smith} and (9) \code{hr}.}

\item{Sigma}{covariance matrix for Brown-Resnick, Smith and extremal student. Default for compatibility}

\item{loc}{matrix of location for Smith model.}

\item{index}{index of the site or variable}

\item{param}{a vector of parameters}
}
\value{
a \code{n} by \code{d} matrix containing the sample
}
\description{
Samples from exceedances at site (scaled extremal function definition)
}
\keyword{internal}
