% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rPSmith}
\alias{.rPSmith}
\title{Generate from Smith model (moving maxima) \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function}
\usage{
.rPSmith(index, Sigma_chol, loc)
}
\arguments{
\item{index}{index of the location. An integer in {0, ..., \eqn{d-1}}}

\item{Sigma_chol}{the Cholesky root of the covariance matrix}

\item{loc}{location matrix}
}
\value{
a \code{d}-vector from \eqn{P_x}
}
\description{
Generate from Smith model (moving maxima) \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function
}
\keyword{internal}
