% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rwdirbsspec}
\alias{.rwdirbsspec}
\title{Generates from \eqn{Q_i}{Qi}, the spectral measure of the weighted Dirichlet model}
\usage{
.rwdirbsspec(n, d, alpha, beta)
}
\arguments{
\item{n}{sample size}

\item{alpha}{vector of concentration parameters}

\item{beta}{vector of Dirichlet components}
}
\value{
a matrix of samples from the angular distribution

an \code{n} by \code{d} sample from the spectral distribution
}
\description{
This model was introduced in Ballani and Schlather (2011).
}
\references{
Ballani, D. and M. Schlather (2011). A construction principle for multivariate extreme value distributions, \emph{Biometrika}, \bold{98}(3), 633--645.
}
\author{
Leo Belzile
}
\keyword{internal}
