% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCdiag.R
\name{.score_algebraic}
\alias{.score_algebraic}
\title{Algebraic score}
\usage{
.score_algebraic(y, x, w, v, m)
}
\arguments{
\item{y}{vector of excesses of lowest threshold \code{u1}}

\item{x}{parameter vector (\code{sigma1}, \code{phi_1}, \ldots, \code{phi_m})}

\item{w}{differences between thresholds (\code{w[m]} not used)}

\item{v}{thresholds relative to lowest threshold}

\item{m}{number of thresholds}
}
\description{
Algebraic score
}
\keyword{internal}
