% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{durbin.unif}
\alias{durbin.unif}
\title{Durbin's mapping of uniforms}
\usage{
durbin.unif(xdat)
}
\arguments{
\item{xdat}{sample of uniform variates}
}
\value{
a data frame with elements \code{stat} containing the test statistics and \code{pval} for the p-values from the null distributions, for the modified probability product, median and (one-sided) Kolmogorov--Smirnov tests.
}
\description{
Given a set of uniform variates, obtained for example through application
of the probability integral  transform, consider alternative scaling which
are also uniform but may lead to more powerful tests.
}
\references{
Durbin, J. (1961). Some Methods of Constructing Exact Tests. \emph{Biometrika}, 48(\bold{1/2}), 41–55. \doi{10.2307/2333128}
}
\keyword{internal}
