% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extgp.R
\name{egp.fitrange}
\alias{egp.fitrange}
\title{Deprecated function for parameter stability plots}
\usage{
egp.fitrange(
  xdat,
  thresh,
  model = c("egp1", "egp2", "egp3"),
  plots = 1:3,
  umin,
  umax,
  nint
)
}
\description{
Deprecated function for parameter stability plots
}
\keyword{internal}
