% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egp.R
\name{egp.rain.fit}
\alias{egp.rain.fit}
\title{Fit an extended generalized Pareto distribution to rainfall intensities}
\usage{
egp.rain.fit(data, model = 1, method = c("ML", "PWM"), init,
  censoring = c(0, Inf), rounded = 0, CI = FALSE, R = 1000, ncpus = 1,
  plots = TRUE)
}
\arguments{
\item{data}{data vector.}

\item{model}{integer ranging from 0 to 4 indicating the model to select}

\item{method}{string; either \code{"ML"} for maximum likelihood, or \code{"PWM"} for probability weighted moments, or both.}

\item{init}{vector of initial values for (\eqn{kappa},\eqn{sigma},\eqn{xi}) for the optimization.}

\item{censoring}{numeric vector of length 2 containing the lower and upper bound for censoring; \code{censoring=c(0,Inf)} is equivalent to no censoring.}

\item{rounded}{numeric giving the instrumental precision (and rounding of the data), with default of 0.}

\item{CI}{logical; should confidence interval be returned (percentile bootstrap).}

\item{R}{integer; number of bootstrap replications.}

\item{ncpus}{integer; number of CPUs for parallel calculations (default: 1).}

\item{plots}{logical; whether to produce histogram and density plots.}
}
\description{
This is a wrapper function to obtain PWM or MLE estimates for
the EGP models of Naveau et al. (2016). The function calculates confidence intervals
by means of nonparametric percentile bootstrap and returns histograms and QQ plots of
the fitted distributions. The function handles both censoring and rounding.
}
\examples{
library(ismev)
data(rain)
egp.rain.fit(rain, model=1, method="ML", init=c(1,gp.fit(rain,0, show=FALSE)$est),rounded=0.1,CI=TRUE)
}
\author{
Raphael Huser
}
\references{
Naveau, P., R. Huser, P. Ribereau, and A. Hannart (2016), Modeling jointly low, moderate, and heavy rainfall intensities without a threshold selection, \emph{Water Resour. Res.}, 52, 2753–2769, \code{doi:10.1002/2015WR018552}.
}
\seealso{
\code{\link[mev]{egp.fit}} \code{\link{egp}}
}

