% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egp.R
\name{pegp2.G}
\alias{pegp2.G}
\alias{degp2.G}
\alias{qegp2.G}
\alias{regp2.G}
\alias{egp2-functions}
\alias{pegp2}
\alias{degp2}
\alias{qegp2}
\alias{regp2}
\title{EGP functions}
\usage{
pegp2.G(u, type = 1, prob, kappa, delta)

degp2.G(u, type = 1, prob = NA, kappa = NA, delta = NA, log = FALSE)

qegp2.G(u, type = 1, prob = NA, kappa = NA, delta = NA)

regp2.G(n, prob = NA, kappa = NA, delta = NA, type = 1,
  unifsamp = NULL, direct = FALSE, censoring = c(0, 1))

pegp2(q, prob = NA, kappa = NA, delta = NA, sigma = NA, xi = NA,
  type = 1)

degp2(x, prob = NA, kappa = NA, delta = NA, sigma = NA, xi = NA,
  type = 1, log = FALSE)

qegp2(p, prob = NA, kappa = NA, delta = NA, sigma = NA, xi = NA,
  type = 1)

regp2(n, prob = NA, kappa = NA, delta = NA, sigma = NA, xi = NA,
  type = 1, unifsamp = NULL, censoring = c(0, Inf))
}
\description{
These functions are documented in \code{\link{egp2}} and in \code{\link{egp2.G}} for the carrier distributions supported in the unit interval.
}
\seealso{
\code{\link{egp2}}, \code{\link{egp2.G}}
}
\keyword{internal}
