% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{eskrain}
\alias{eskrain}
\title{Eskdalemuir Observatory Daily Rainfall}
\format{
a vector with 93 daily cumulated rainfall measurements exceeding 30mm.
}
\source{
Met Office.
}
\description{
This dataset contains exceedances of 30mm for daily
cumulated rainfall observations over the period 1970-1986.
These data were aggregated from hourly series.
}
\details{
The station is one of the rainiest of the whole UK, with an average 1554m of cumulated rainfall per year.
The data consisted of 6209 daily observations, of which 4409 were non-zero.
Only the 93 largest observations are provided.
}
