% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expme_rparpcs.R
\name{expme}
\alias{expme}
\title{Exponent measure for elliptical extreme value processes}
\usage{
expme(z, Sigma = NULL, Lambda = NULL, m = NULL, df = NULL,
  model = c("hr", "br", "xstud"), method = c("mvPot", "TruncatedNormal",
  "mvtnorm"))
}
\arguments{
\item{z}{vector at which to estimate exponent measure}

\item{Sigma}{correlation matrix if \code{model = "xstud"}, otherwise
the covariance matrix formed from the stationary Brown-Resnick process.}

\item{Lambda}{parameter matrix for the Brown--Resnick model}

\item{m}{mean vector for the Huesler-Reiss model}

\item{df}{degrees of freedom for extremal Student process}

\item{model}{string indicating the model family}

\item{method}{string indicating the package from which to extract the numerical integration routine}
}
\value{
numeric
}
\description{
Measure of the set \eqn{[0, z]^c} for Huesler-Reiss, Brown-Resnick and extremal Student processes.
The expression appears in likelihood inference
}
\examples{
#Extremal Student
Sigma <- stats::rWishart(n = 1, df = 20, Sigma = diag(10))[,,1]
expme(z = rep(1, ncol(Sigma)), Sigma = cov2cor(Sigma), df = 3, model = "xstud")
#Brown-Resnick model
D <- 5L
loc <- cbind(runif(D), runif(D))
di <- as.matrix(dist(rbind(c(0, ncol(loc)), loc)))
semivario <- function(d, alpha = 1.5, lambda = 1){(d/lambda)^alpha}
Vmat <- semivario(di)
Lambda <- Vmat[-1,-1]/2
expme(z <- rep(1, ncol(Lambda)), Lambda = Lambda, model = "br", method = "mvPot")
Sigma <- outer(Vmat[-1, 1], Vmat[1, -1], "+") - Vmat[-1, -1]
expme(z <- rep(1, ncol(Lambda)), Lambda = Lambda, model = "br", method = "mvPot")
}
