% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extcoef.R
\name{extcoef}
\alias{extcoef}
\title{Estimator of the extremal coefficient}
\usage{
extcoef(dat, loc, thresh, estimator = c("schlather", "fmado"),
  prob = 0, which.plot = c("schlather", "fmado"), tikz = FALSE)
}
\arguments{
\item{dat}{an \code{n} by \code{D} matrix of unit Frechet observations}

\item{loc}{\code{d} by \code{D} matrix of location coordinates}

\item{thresh}{threshold parameter (dafault is to keep all data if \code{prob = 0}.}

\item{estimator}{string indicating which model estimates to compute, one of \code{schlather} or \code{fmado}.}

\item{prob}{probability of not exceeding threshold \code{thresh}}

\item{which.plot}{string indicating which model estimates to plot, one of \code{schlather} or \code{fmado}. Can be abbreviated.}

\item{tikz}{logical indicating whether output be formatted to use with \code{tikzDevice} package? Default to \code{FALSE}.}
}
\value{
an invisible list with components \code{dist}, \code{schlather}, \code{fmado} and \code{binned}. The first three are \code{n} vectors, while \code{binned} is a matrix with 2 columns containing the binned distance \code{h} and the binned extremal coefficient.
}
\description{
This function returns a matrix with the Schlather and Tawn estimator of extremal coefficient
based on the extremal index of pairs of max-stable vectors. The likelihood of the naive estimator for a pair of two sites \eqn{A} is
\deqn{ \mathrm{card}\left\{ j: \max_{i \in A} X_i^{(j)}\bar{X}_i)>z \right\} \log(\theta_A) - \theta_A \sum_{j=1}^n \left[ \max \left\{z, \max_{i \in A} (X_i^{(j)}\bar{X}_i)\right\}\right]^{-1},}
where \eqn{\bar{X}_i = n^{-1} \sum_{j=1}^n 1/X_i^{(j)}} is the harmonic mean and \eqn{z} is a threshold on the unit Frechet scale.
The search for the maximum likelihood estimate for every pair \eqn{A} is restricted to the interval \eqn{[1,2]}.
The Schlather estimator is not self-consistent. A binned version of the extremal coefficient cloud is also reported and superimposed on the graph.
}
\details{
Additionally, a poor man's version of the F-madogram estimator is returned alongside with the Schlather and Tawn estimates.
If \code{estimator = c('schlather', 'fmado')} and \code{plot = TRUE}, both data clouds are plotted alongside one another.

The F-madogram estimator is
\deqn{\nu(h) = \frac{1}{2} \mathsf{E} |F(Z(x+h))-F(Z(x))|.}
We can write the extremal coefficient as
\deqn{\theta(h) = \frac{1+2\nu(h)}{1-2\nu(h)},}
leading to a non-parametric pairwise estimator. The empirical cloud of estimates is obtained by using a rank-based
approximation to the distribution function \eqn{F}.

The implementation only uses complete pairs to calculate the relative ranks. Both estimators are coded in plain R and computations are not optimized. The estimation time can therefore be significant for large data sets.
}
\examples{
\dontrun{
loc <- 10*cbind(runif(50), runif(50))
di <- as.matrix(dist(loc))
dat <- rmev(n = 1000, d = 100, param = 3, sigma = exp(-di/2), model = 'xstud')
res <- extcoef(dat = dat, loc = loc)
Extremal Student extremal coefficient function

XT.extcoeffun <- function(h, nu, corrfun, ...){
  if(!is.function(corrfun)){
    stop('Invalid function `corrfun`.')
  }
  h <- unique(as.vector(h))
  rhoh <- sapply(h, corrfun, ...)
  cbind(h = h, extcoef = 2*pt(sqrt((nu+1)*(1-rhoh)/(1+rhoh)), nu+1))
}
#This time, only one graph with theoretical extremal coef
plot(res$dist, res$schlather, ylim = c(1,2), pch = 20); abline(v = 2, col = 'gray')
extcoefxt <- XT.extcoeffun(seq(0, 10, by = 0.1), nu = 3,
                            corrfun = function(x){exp(-x/2)})
lines(extcoefxt[,'h'], extcoefxt[,'extcoef'], type = 'l', col = 'blue', lwd = 2)
# Brown--Resnick extremal coefficient function
BR.extcoeffun <- function(h, vario, ...){
  if(!is.function(vario)){
    stop('Invalid function `vario`.')
  }
  h <- unique(as.vector(h))
  gammah <- sapply(h, vario, ...)
  cbind(h = h, extcoef = 2*pnorm(sqrt(gammah/4)))
}
extcoefbr<- BR.extcoeffun(seq(0, 20, by = 0.25), vario = function(x){2*x^0.7})
lines(extcoefbr[,'h'], extcoefbr[,'extcoef'], type = 'l', col = 'orange', lwd = 2)
}
}
\references{
Schlather, M. and J. Tawn (2003). A dependence measure for multivariate and spatial extremes, \emph{Biometrika}, \bold{90}(1), pp.139--156.

Cooley, D., P. Naveau and P. Poncet (2006). Variograms for spatial max-stable random fields,  In: Bertail P., Soulier P., Doukhan P. (eds) \emph{Dependence in Probability and Statistics}. Lecture Notes in Statistics, vol. 187. Springer, New York, NY
}
