% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extcoef.R
\name{extcoef.smith}
\alias{extcoef.smith}
\title{Pairwise extremal coefficient estimate of Smith}
\usage{
extcoef.smith(dat, loc = NULL, standardize = TRUE,
  method = c("nonparametric", "parametric"))
}
\arguments{
\item{dat}{an \code{n} by \code{d} matrix of maximas for \code{d} variables}

\item{loc}{\code{d} by \code{D} matrix of coordinates; default to \code{NULL}}

\item{standardize}{logical; should observations be transformed to unit Frechet scale? Default is to transform}
}
\description{
Suppose \eqn{Z(x)} is  a max-stable vector with unit Frechet distribution. Then
 \eqn{1/Z} is unit exponential and \eqn{1/\max(Z(s_1), Z(s_2))} is exponential
 with rate \eqn{\theta = \max(Z(s_1), Z(s_2))}.
 The extremal index for the pair can therefore be calculated using the reciprocal mean.
}
\examples{
loc <- 10*cbind(runif(20), runif(20))
di <- as.matrix(dist(loc))
dat <- rmev(n = 1000, d = 20, param = 3, sigma = exp(-di/2), model = 'xstud')
res <- extcoef.smith(dat = dat, loc = loc)
}
\references{
R. J. Erhardt, R. L. Smith (2012), Approximate Bayesian computing for spatial extremes, \emph{Computational Statistics and Data Analysis}, \bold{56}, pp.1468--1481.
}
