% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extcoef.R
\name{extcoefschlather}
\alias{extcoefschlather}
\title{Schlather's estimator of the extremal coefficient}
\usage{
extcoefschlather(dat, loc, thresh, estimator = c("schlather", "fmado"),
  prob = 0, which.plot = c("schlather", "fmado"), tikz = FALSE)
}
\arguments{
\item{dat}{a matrix of unit Fréchet observations}

\item{thresh}{threshold parameter}

\item{estimator}{string indicating which model estimates to compute, one of \code{schlather} or \code{fmado}.}

\item{prob}{probability of not exceeding threshold \code{thresh}}

\item{which.plot}{string indicating which model estimates to plot, one of \code{schlather} or \code{fmado}. Can be abbreviated.}

\item{tikz}{logical indicating whether output be formatted to use with \code{tikzDevice} package? Default to \code{FALSE}.}
}
\description{
This functions returns a matrix with the Schlather and Tawn estimator of extremal coefficient
based on the extremal index of pairs of max-stable vectors. The likelihood of the naive estimator for a pair of two sites \eqn{A} is
\deqn{ \mathrm{card}\left\{ j: \max_{i \in A} X_i^{(j)}\bar{X}_i)>z \right\} \log(\theta_A) - \theta_A \sum_{j=1}^n \left[ \max \left\{z, \max_{i \in A} (X_i^{(j)}\bar{X}_i)\right\}\right]^{-1},}
where \eqn{\bar{X}_i = n^{-1} \sum_{j=1}^n 1/X_i^{(j)}} is the harmonic mean and \eqn{z} is a threshold on the unit Frechet scale.
The search for the maximum likelihood estimate for every pair \eqn{A} is restricted to the interval \eqn{[1,2]}.
The Schlather estimator is not self-consistent. A binned version of the extremal coefficient cloud is also reported and superimposed on the graph.
}
\details{
The F-madogram estimator is
\deqn{\nu(h) = \frac{1}{2} \mathsf{E} |F(Z(x+h))-F(Z(x))|.}
We can write the extremal coefficient as
\deqn{\theta(h) = \frac{1+2\nu(h)}{1-2\nu(h)},}
leading to a non-parametric pairwise estimator. The empirical cloud of estimates is obtained by using a rank-based
approximation to the distribution function \eqn{F}.

The implementation only uses complete pairs to calculate the relative ranks. Both estimators are coded in plain R and
the estimation time can be significant.
}
\examples{
loc <- 10*cbind(runif(50), runif(50))
di <- as.matrix(dist(loc))
dat <- mev::rmev(n = 1000, d = 100, param = 3, sigma = exp(-di/2), model = "xstud")
extcoefschlather(dat = dat, loc = loc)
}
\references{
Schlather, M. and J. Tawn (2003). A dependence measure for multivariate and spatial extremes, \emph{Biometrika}, \bold{90}(1), pp.139--156.

Cooley, D., P. Naveau and P. Poncet (2006). Variograms for spatial max-stable random fields,  In: Bertail P., Soulier P., Doukhan P. (eds) \emph{Dependence in Probability and Statistics}. Lecture Notes in Statistics, vol. 187. Springer, New York, NY
}
