% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condextr.R
\name{fit.condex}
\alias{fit.condex}
\title{Pseudo maximum likelihood estimation for the conditional extremes model}
\usage{
fit.condex(
  xdat,
  thresh,
  mthresh,
  index = 1L,
  ties.method = c("average", "first", "last", "random", "max", "min"),
  marginal = c("empirical", "none", "semipar"),
  use = c("complete", "pairwise"),
  constraint = TRUE,
  ...
)
}
\description{
Numerical optimization of the Gaussian pseudo likelihood for the conditional extremes model of Heffernan and Tawn, with constraints.
}
