% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evdmle.R
\name{fit.gev}
\alias{fit.gev}
\title{Maximum likelihood estimation for the generalized extreme value distribution}
\usage{
fit.gev(xdat, show = FALSE)
}
\arguments{
\item{xdat}{a numeric vector of data to be fitted.}

\item{show}{logical; if \code{TRUE} (the default), print details of the fit.}
}
\value{
a list containing the following components:
\itemize{
\item \code{estimate} a vector containing all parameters.
\item \code{std.err} a vector containing the standard errors.
\item \code{var.cov} the variance covariance matrix, obtained as the numerical inverse of the observed information matrix.
\item \code{method} the method used to fit the parameter.
\item \code{deviance} the deviance at the maximum likelihood estimates.
\item \code{convergence} components taken from the list returned by \code{\link[alabama]{constrOptim.nl}}.
Values other than \code{0} indicate that the algorithm likely did not converge.
\item \code{counts} components taken from the list returned by \code{\link[alabama]{constrOptim.nl}}.
}
}
\description{
This function returns an object of class \code{gev}, with default methods for printing and quantile-quantile plots.
}
