% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{fit.gevblock.marginal}
\alias{fit.gevblock.marginal}
\title{Optimization for the marginal likelihood for a sample of GEV order statistics}
\usage{
fit.gevblock.marginal(
  xdat,
  constraint = TRUE,
  rounding = 0,
  lb = NULL,
  start = NULL,
  vcov = FALSE
)
}
\arguments{
\item{xdat}{matrix of observations of size \code{n} by \code{m}, ordered by rows}

\item{constraint}{logical; if \code{TRUE}, add support constraint}

\item{rounding}{double; indicate the amount of rounding around value; default to zero}

\item{lb}{lower bound; any point below \code{lb} is left-censored}

\item{start}{vector of length 3 for starting values for GEV; default to \code{NULL}}

\item{vcov}{logical; if \code{TRUE}, return as attribute the estimate of the covariance matrix of the parameters given by the inverse observed information matrix.}
}
\value{
(constrained) maximum likelihood estimator of location, scale and shape parameters
}
\description{
Given a matrix of \code{n} ordered samples of \code{m} order statistics from a postulated GEV, fit the parameters of the latter based on the marginal likelihood of the first
\code{m-1} order statistics using maximum likelihood.
}
\details{
Additionally, one can set \code{constraint} to \code{TRUE} to add a support constraint
to the optimization to ensure that all values of \code{xdat} are in the support of the resulting distribution.
}
\keyword{internal}
