% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{fit.gevblock}
\alias{fit.gevblock}
\title{Optimization for the GEV likelihood for blocks}
\usage{
fit.gevblock(
  xdat,
  marginal = FALSE,
  constraint = TRUE,
  rounding = 0,
  lb = NULL,
  start = NULL,
  vcov = FALSE
)
}
\arguments{
\item{xdat}{matrix of observations of size \code{n} by \code{m}, ordered by rows}

\item{marginal}{logical; if \code{TRUE}, use marginal likelihood of lower order statistics}

\item{constraint}{logical; if \code{TRUE}, add support constraint}

\item{rounding}{double; indicate the amount of rounding around value; default to zero}

\item{lb}{lower bound; any point below \code{lb} is left-censored}

\item{start}{vector of length 3 for starting values for GEV; default to \code{NULL}}

\item{vcov}{logical; if \code{TRUE}, return as attribute the estimate of the covariance matrix of the parameters given by the inverse observed information matrix.}
}
\value{
(constrained) maximum likelihood estimator of location, scale and shape parameters
}
\description{
Given a matrix of \code{n} ordered samples of \code{m} order statistics
from a postulated GEV, fit the parameters of the latter based on the
 marginal likelihood of the first \code{m-1} order statistics using
 maximum likelihood.
}
\details{
One can set \code{constraint} to \code{TRUE} to add a support constraint
to the optimization to ensure that all values of \code{xdat} are in the
support of the resulting distribution (only for the marginal likelihood).
}
\examples{
set.seed(2026)
xdat <- build.blocks(mev::rgev(n = 200, shape = 0.1), m = 4)
fit.gevblock(xdat, marginal = TRUE)
fit.gevblock(round(xdat, 1), marginal = TRUE, lb = NULL, rounding = 0.1)
fit.gevblock(round(xdat, 1), marginal = TRUE, lb = -2, rounding = 0.1)
fit.gevblock(xdat, marginal = TRUE, lb = -2)
fit.gevblock(xdat)
fit.gevblock(round(xdat, 1), lb = NULL, rounding = 0.1)
fit.gevblock(round(xdat, 1), lb = -2, rounding = 0.1)
fit.gevblock(xdat, lb = -2)
}
