% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{fit.rho}
\alias{fit.rho}
\title{Estimator of the second order tail index parameter}
\usage{
fit.rho(xdat, k, method = c("fagh", "dk", "ghp"), ...)
}
\arguments{
\item{xdat}{vector of positive observations}

\item{k}{number of highest order statistics to use for estimation}

\item{method}{string for the estimator}

\item{...}{additional arguments passed to individual routinescurrently ignored.}
}
\description{
Estimator of the second order tail index parameter
}
