% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{frwind}
\alias{frwind}
\title{French wind data}
\format{

}
\source{
European Climate Assessment and Dataset project \url{https://www.ecad.eu/}
}
\description{
Daily mean wind speed (in km/h) at four stations in the south of France, namely Cap Cépet (\code{S1}), Lyon St-Exupéry (\code{S2}), Marseille Marignane (\code{S3}) and Montélimar (\code{S4}).
The data includes observations from January 1976 until April 2023; days containing missing values are omitted.
}
\examples{
data(frwind, package = "mev")
head(frwind)
attr(frwind, which = "metadata")
}
\references{
Klein Tank, A.M.G. and Coauthors, 2002. Daily dataset of 20th-century surface air temperature and precipitation series for the
European Climate Assessment. Int. J. of Climatol., 22, 1441-1453.
}
