% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egp.R
\name{pG}
\alias{dG}
\alias{pG}
\alias{qG}
\alias{rG}
\title{Carrier distribution for extended generalized Pareto distributions}
\usage{
pG(u, type = 1, prob, kappa, delta)

dG(u, type = 1, prob = NA, kappa = NA, delta = NA, log = FALSE)

qG(u, type = 1, prob = NA, kappa = NA, delta = NA)

rG(n, prob = NA, kappa = NA, delta = NA, type = 1, unifsamp = NULL,
  direct = FALSE, censoring = c(0, 1))
}
\arguments{
\item{u}{vector of observations (\code{dG}), probabilities (\code{qG}) or quantiles (\code{pG}), in \eqn{[0,1]}}

\item{type}{integer between 0 to 5 giving the model choice}

\item{prob}{mixture probability for model \code{type} \code{4}}

\item{kappa}{shape parameter for \code{type} \code{1} ,\code{3} and \code{4}}

\item{delta}{additional parameter for \code{type} \code{2} ,\code{3} and \code{4}}

\item{log}{logical; should the log-density be returned (default to \code{FALSE})?}

\item{n}{sample size}

\item{unifsamp}{sample of uniform; if provided, the data will be used in place of new uniform random variates}

\item{direct}{logical; which method to use for sampling in model of \code{type} \code{4}?}

\item{censoring}{numeric vector of length 2 containing the lower and upper bound for censoring}
}
\description{
Density, distribution function, quantile function and random number
generation for the carrier distributions of the extended Generalized Pareto distributions.
}
\author{
Raphael Huser
}
\seealso{
\code{\link{egp}}
}

