% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{generate.exponential}
\alias{generate.exponential}
\title{Generate exponential}
\usage{
generate.exponential(n, m)
}
\arguments{
\item{n}{number of rows}

\item{m}{number of columns}
}
\value{
an \code{n} by \code{m} matrix of order statistics from unit exponential
}
\description{
Given integers \code{n} and \code{m}, simulate a matrix of
independent unit exponential order statistics,
but ordered by row from largest to smallest
}
\keyword{internal}
