% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{gev.abias}
\alias{gev.abias}
\title{Asymptotic bias of block maxima for fixed sample sizes}
\usage{
gev.abias(shape, rho)
}
\arguments{
\item{shape}{shape parameter}

\item{rho}{second-order parameter, non-positive}
}
\value{
a vector of length three containing the bias for location, scale and shape (in this order)
}
\description{
Asymptotic bias of block maxima for fixed sample sizes
}
\references{
Dombry, C. and A. Ferreira (2017). Maximum likelihood estimators based on the block maxima method. \code{https://arxiv.org/abs/1705.00465}
}
