% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{gev.Fscore}
\alias{gev.Fscore}
\title{Firth's modified score equation for the generalized extreme value distribution}
\usage{
gev.Fscore(par, dat, method = "obs")
}
\arguments{
\item{par}{vector of \code{loc}, \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{method}{string indicating whether to use the expected ('exp') or the observed ('obs' - the default) information matrix.}
}
\description{
Firth's modified score equation for the generalized extreme value distribution
}
\references{
Firth, D. (1993). Bias reduction of maximum likelihood estimates, \emph{Biometrika}, \strong{80}(1), 27--38.
}
\seealso{
\code{\link{gev}}
}
\keyword{internal}
