% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gev.ll}
\alias{gev.ll}
\alias{gev.ll.optim}
\title{log likelihood for the generalized extreme value distribution}
\usage{
gev.ll(par, dat)

gev.ll.optim(par, dat)
}
\arguments{
\item{par}{vector of \code{loc}, \code{scale} and \code{shape}}

\item{dat}{sample vector}
}
\description{
Function returning the density of an \code{n} sample from the GEV distribution.
}
\details{
\code{gev.ll.optim} returns the negative log likelihood parametrized in terms of location, \code{log(scale)} and shape in order to perform unconstrained optimization
}
\seealso{
\code{\link{gev}}
}
\keyword{internal}
