% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gev.retlev}
\alias{gev.retlev}
\title{Return level for the generalized extreme value distribution}
\usage{
gev.retlev(par, p)
}
\arguments{
\item{par}{vector of \code{loc}, \code{scale} and \code{shape}}

\item{p}{vector of probabilities}
}
\description{
This function returns the \eqn{1-p}th quantile of the GEV.
}
\seealso{
\code{\link{gev}}
}
\keyword{internal}
