% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{gev.score}
\alias{gev.score}
\title{Score vector for the generalized extreme value distribution}
\usage{
gev.score(par, dat)
}
\arguments{
\item{par}{vector of \code{loc}, \code{scale} and \code{shape}}

\item{dat}{sample vector}
}
\description{
Score vector for the generalized extreme value distribution
}
\keyword{internal}

