% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gev.temstat}
\alias{gev.temstat}
\alias{gev.Vfun}
\alias{gev.phi}
\alias{gev.dphi}
\title{Tangent exponential model statistics for the generalized extreme value distribution}
\usage{
gev.Vfun(par, dat)

gev.phi(par, dat, V)

gev.dphi(par, dat, V)
}
\arguments{
\item{par}{vector of \code{loc}, \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{V}{vector calculated by \code{gev.Vfun}}
}
\description{
Matrix of approximate ancillary statistics, sample space derivative of the
log likelihood and mixed derivative for the generalized extreme value distribution.
}
\seealso{
\code{\link{gev}}
}
\keyword{internal}
