% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{gev2exp}
\alias{gev2exp}
\title{Transform order statistics of GEV to exponential}
\usage{
gev2exp(xdat, par)
}
\arguments{
\item{xdat}{matrix of observations (ordered by row) from GEV}

\item{par}{vector of location, scale and shape parameters of the GEV}
}
\value{
a matrix of the same dimension as \code{xdat} containing the
ordered unit exponential quantiles
}
\description{
Transform order statistics of GEV to exponential
}
\keyword{internal}
