% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{gevblock.ll}
\alias{gevblock.ll}
\title{Generalized extreme value model by block}
\usage{
gevblock.ll(
  pars,
  xdat,
  null = FALSE,
  lb = NULL,
  rounding = 0,
  alternative = 1:3,
  logscale = FALSE
)
}
\arguments{
\item{pars}{vector of location, scale and shape parameters for the GEV}

\item{xdat}{\code{n} by \code{m} matrix of observations assumed to arise from a GEV, ordered by row}

\item{null}{logical; if \code{TRUE}, the parameter vector for the largest observation of the block is obtained from the max stability relationship}

\item{lb}{lower bound for left-censoring; default to none (\code{NULL})}

\item{rounding}{double, a positive number indicating the amount of censoring (e.g., 0.1 or 1)}

\item{alternative}{integer; \code{1} for a single shape parameter with parameters obtained from max-stability, \code{2} for a common shape parameter, but with free location and scale, and \code{3} for different parameters (location, scale and shape).}

\item{logscale}{logical; if \code{TRUE}, scale parameter is provided on the log scale}
}
\value{
log likelihood of sample
}
\description{
Given a sample of block maxima of dimension \code{n} by \code{m}, split further into groups of size \code{m} (number of columns), fit a generalized extreme value distribution allowing the maximum of each subblock of size \code{m} to have different parameters under the alternative. Likelihood for left-censored and rounded (interval-censored) observations is
also supported.
}
\examples{
samp <- build.blocks(mev::rgev(50, scale = 10), m = 4)
pars <- c(0,10,0)
gevblock.ll(pars, xdat = round(samp,0), null = TRUE, rounding = 1, lb = -5)
gevblock.ll(pars, xdat = round(samp,0), null = TRUE, rounding = 1)
gevblock.ll(pars, xdat = samp, null = TRUE)
gevblock.ll(pars, xdat = samp, lb = -5, null = TRUE)
}
\keyword{internal}
