% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gevN.infomat}
\alias{gevN.infomat}
\title{Information matrix of the generalized extreme value distribution (quantile/mean of N-block maxima parametrization)}
\usage{
gevN.infomat(
  par,
  dat,
  method = c("obs", "exp"),
  qty = c("mean", "quantile"),
  N,
  q = 0.5,
  nobs = length(dat)
)
}
\arguments{
\item{par}{vector of \code{loc}, quantile/mean of N-block maximum and \code{shape}}

\item{dat}{sample vector}

\item{qty}{string indicating whether to calculate the \code{q} quantile or the mean}

\item{q}{probability, corresponding to \eqn{q}th quantile of the \code{N}-block maximum}
}
\description{
Information matrix of the generalized extreme value distribution (quantile/mean of N-block maxima parametrization)
}
\seealso{
\code{\link{gevN}}
}
\keyword{internal}
