% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gevN.ll}
\alias{gevN.ll}
\title{Negative log likelihood of the generalized extreme value distribution (quantile/mean of N-block maxima parametrization)}
\usage{
gevN.ll(par, dat, N, q = 0.5, qty = c("mean", "quantile"))
}
\arguments{
\item{par}{vector of \code{loc}, quantile/mean of N-block maximum and \code{shape}}

\item{dat}{sample vector}

\item{q}{probability, corresponding to \eqn{q}th quantile of the \code{N}-block maximum}

\item{qty}{string indicating whether to calculate the \code{q} quantile or the mean}
}
\description{
Negative log likelihood of the generalized extreme value distribution (quantile/mean of N-block maxima parametrization)
}
\seealso{
\code{\link{gevN}}
}
\keyword{internal}
