% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gevN.mean}
\alias{gevN.mean}
\title{This function returns the mean of N observations from the GEV.}
\usage{
gevN.mean(par, N)
}
\arguments{
\item{par}{vector of \code{loc}, quantile/mean of N-block maximum and \code{shape}}
}
\description{
This function returns the mean of N observations from the GEV.
}
\seealso{
\code{\link{gevN}}
}
\keyword{internal}
