% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{gevr.infomat}
\alias{gevr.infomat}
\title{Observed information matrix for GEV distribution (return levels)}
\usage{
gevr.infomat(par, dat, p)
}
\arguments{
\item{par}{vector of \code{retlev}, \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{p}{probability, corresponding to \eqn{(1-p)}th quantile for \eqn{z}}
}
\description{
The information matrix is parametrized in terms of location, \eqn{(1-p)}th quantile and shape.
}
\seealso{
\code{\link{gevr}}
}
\keyword{internal}

