% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{gevr}
\alias{gevr}
\alias{gevr.Vfun}
\alias{gevr.dphi}
\alias{gevr.infomat}
\alias{gevr.ll}
\alias{gevr.ll.optim}
\alias{gevr.phi}
\alias{gevr.score}
\title{Generalized extreme value distribution (return level parametrization)}
\usage{
gevr.ll(par, dat, p)

gevr.ll.optim(par, dat, p)

gevr.score(par, dat, p)

gevr.infomat(par, dat, p)

gevr.Vfun(par, dat, p)

gevr.phi(par, dat, p, V)

gevr.dphi(par, dat, p, V)
}
\arguments{
\item{par}{vector of parameters, \code{loc}, \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{p}{probability, corresponding to (1-p)th quantile for \eqn{z}}

\item{V}{vector calculated by \code{\link{gev.Vfun}}}
}
\description{
Likelihood, score function and information matrix, bias, approximate ancillary statistics and sample space derivative for the generalized extreme value distribution parametrized in terms of the return level \eqn{z}, scale and shape.
}
\section{Functions}{
\itemize{
\item \code{gevr.ll}: log-likelihood

\item \code{gevr.ll.optim}: log-likelihood

\item \code{gevr.score}: log-likelihood

\item \code{gevr.infomat}: observed information matrix

\item \code{gevr.Vfun}: log-likelihood

\item \code{gevr.phi}: canonical parameter in the local exponential family approximation

\item \code{gevr.dphi}: derivative matrix of the canonical parameter in the local exponential family approximation
}}
\seealso{
\code{\link{gev}}
}

