% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{gevr.score}
\alias{gevr.score}
\title{Score of the log-likelihood for the GEV distribution (return levels)}
\usage{
gevr.score(par, dat, p)
}
\arguments{
\item{par}{vector of \code{retlev}, \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{p}{probability, corresponding to \eqn{(1-p)}th quantile for \eqn{z}}
}
\description{
Score of the log-likelihood for the GEV distribution (return levels)
}
\seealso{
\code{\link{gevr}}
}
\keyword{internal}

