% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{gevr.temstat}
\alias{gevr.Vfun}
\alias{gevr.dphi}
\alias{gevr.phi}
\alias{gevr.temstat}
\title{Tangent exponential model statistics for the GEV distribution (return level)}
\usage{
gevr.Vfun(par, dat, p)

gevr.phi(par, dat, p, V)

gevr.dphi(par, dat, p, V)
}
\arguments{
\item{par}{vector of \code{retlev}, \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{p}{probability, corresponding to \eqn{(1-p)}th quantile for \eqn{z}}

\item{V}{vector calculated by \code{gevr.Vfun}}
}
\description{
Vector implementing conditioning on approximate ancillary statistics for the TEM

Canonical parameter in the local exponential family approximation

Derivative of the canonical parameter \eqn{\phi(\theta)} in the local exponential family approximation
}
\seealso{
\code{\link{gevr}}
}
\keyword{internal}

