% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gevrl.ll}
\alias{gevrl.ll}
\title{Log-likelihood of the point process of r-largest observations}
\usage{
gevrl.ll(par, dat)
}
\arguments{
\item{par}{a vector of location, scale and shape parameter}

\item{dat}{an \code{n} by \code{r} matrix of observations}
}
\description{
Log-likelihood of the point process of r-largest observations
}
\keyword{internal}
