% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gevrl.score}
\alias{gevrl.score}
\title{Score of the r-largest observations}
\usage{
gevrl.score(par, dat)
}
\arguments{
\item{par}{a vector of location, scale and shape parameter}

\item{dat}{an \code{n} by \code{r} matrix of observations}
}
\value{
a vector of size 3
}
\description{
The score is computed via linear interpolation for the shape parameter in a neighborhood of zero
}
\keyword{internal}
