% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{gpd.abias}
\alias{gpd.abias}
\title{Asymptotic bias of threshold exceedances for k order statistics}
\usage{
gpd.abias(shape, rho)
}
\arguments{
\item{shape}{shape parameter}

\item{rho}{second-order parameter, non-positive}
}
\value{
a vector of length containing the bias for scale and shape (in this order)
}
\description{
The formula given in de Haan and Ferreira, 2007 (Springer). Note that the latter differs from that found in Drees, Ferreira and de Haan.
}
\references{
Dombry, C. and A. Ferreira (2017). Maximum likelihood estimators based on the block maxima method. \code{https://arxiv.org/abs/1705.00465}
}
