% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{gpd.Fscore}
\alias{gpd.Fscore}
\title{Firth's modified score equation for the generalized Pareto distribution}
\usage{
gpd.Fscore(par, dat, method = c("obs", "exp"))
}
\arguments{
\item{par}{vector of \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{method}{string indicating whether to use the expected  (\code{"exp"}) or the observed (\code{"obs"} - the default) information matrix.}
}
\description{
Firth's modified score equation for the generalized Pareto distribution
}
\references{
Firth, D. (1993). Bias reduction of maximum likelihood estimates, \emph{Biometrika}, \strong{80}(1), 27--38.
}
\seealso{
\code{\link{gpd}}, \code{\link{gpd.bcor}}
}
\keyword{internal}

