% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm.R
\name{gpd.lmom}
\alias{gpd.lmom}
\title{Estimation of generalized Pareto parameters via L-moments}
\usage{
gpd.lmom(xdat, thresh, sorted = FALSE, Lskew = FALSE)
}
\arguments{
\item{xdat}{[numeric] vector of observations}

\item{thresh}{[numeric] optional threshold argument}

\item{sorted}{[logical] if \code{TRUE}, observations are sorted in increasing order}

\item{Lskew}{[logical]; if \code{TRUE}, shape is obtained from L-skewness rather than first two moments.}

\item{return}{string, either \code{vector} for a vector of length 2 from the parameters, or \code{list} for an object of class \code{mev_gpd}}
}
\value{
a vector of length two with the scale and shape estimates
}
\description{
Given a sample of exceedances, compute the first four L-moments and use either the first two to obtain the scale and shape (default), or else use L-skewness and L-scale to compute the scale and shape of the generalized Pareto distribution
}
