% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpd.score}
\alias{gpd.score}
\title{Score vector for the generalized Pareto distribution}
\usage{
gpd.score(par, dat)
}
\arguments{
\item{par}{vector of \code{scale} and \code{shape}}

\item{dat}{sample vector}
}
\description{
Score vector for the generalized Pareto distribution
}
\seealso{
\code{\link{gpd}}
}
\keyword{internal}
