% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{gpd.temstat}
\alias{gpd.Vfun}
\alias{gpd.dphi}
\alias{gpd.phi}
\alias{gpd.temstat}
\title{Tangent exponential model statistics for the generalized Pareto distribution}
\usage{
gpd.Vfun(par, dat)

gpd.phi(par, dat, V)

gpd.dphi(par, dat, V)
}
\arguments{
\item{par}{vector of \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{V}{vector calculated by \code{gpd.Vfun}}
}
\description{
Matrix of approximate ancillary statistics, sample space derivative of the
log-likelihood and mixed derivative for the generalized Pareto distribution.
}
\seealso{
\code{\link{gpd}}
}
\keyword{internal}

