% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{gpde.infomat}
\alias{gpde.infomat}
\title{Observed information matrix for the GP distribution (expected shortfall)}
\usage{
gpde.infomat(par, dat, m)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{y_m} and \eqn{\xi}, respectively the \eqn{m}-year return level and the shape parameter.}

\item{dat}{sample vector}

\item{m}{number of observations of interest for return levels. See \strong{Details}}
}
\description{
The information matrix is parametrized in terms of rate of expected shortfall and shape
}
\seealso{
\code{\link{gpde}}
}
\keyword{internal}

