% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{gpde}
\alias{gpde}
\alias{gpde.Vfun}
\alias{gpde.dphi}
\alias{gpde.infomat}
\alias{gpde.ll}
\alias{gpde.ll.optim}
\alias{gpde.phi}
\alias{gpde.score}
\title{Generalized Pareto distribution (expected shortfall parametrization)}
\usage{
gpde.ll(par, dat, m)

gpde.ll.optim(par, dat, m)

gpde.score(par, dat, m)

gpde.infomat(par, dat, m)

gpde.Vfun(par, dat, m)

gpde.phi(par, dat, V, m)

gpde.dphi(par, dat, V, m)
}
\arguments{
\item{par}{vector of parameters (expected shortfall and shape)}

\item{dat}{sample vector.}

\item{m}{number of observations of interest for return levels. See \strong{Details}}

\item{V}{vector implementing conditioning on approximate ancillary statistics for the TEM}
}
\description{
Likelihood, score function and information matrix, bias,
approximate ancillary statistics and sample space derivative
for the generalized Pareto distribution (expected shortfall parametrization)
}
\details{
The observed information matrix was calculated using the second Bartlett identity as the negative of the second derivative of the log-likelihood function, using symbolic calculus in Sage.

The interpretation for \code{m} is as follows: if there are on average \eqn{m_y} observations per year above the threshold, then  \eqn{m=Tm_y} corresponds to \eqn{T}-year return level.
}
\section{Functions}{
\itemize{
\item \code{gpde.ll}: log-likelihood

\item \code{gpde.ll.optim}: negative log-likelihood parametrized in terms of log expected shortfall and shape in order to perform unconstrained optimization

\item \code{gpde.score}: score vector

\item \code{gpde.infomat}: observed information matrix for GPD parametrized in terms of rate of expected shortfall and shape

\item \code{gpde.Vfun}: vector implementing conditioning on approximate ancillary statistics for the TEM

\item \code{gpde.phi}: canonical parameter in the local exponential family approximation

\item \code{gpde.dphi}: derivative matrix of the canonical parameter in the local exponential family approximation
}}
\author{
Leo Belzile
}
\seealso{
\code{\link{gpd}}
}

