% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{gpde.temstat}
\alias{gpde.Vfun}
\alias{gpde.dphi}
\alias{gpde.phi}
\alias{gpde.temstat}
\title{Tangent exponential model statistics for the generalized Pareto distribution (expected shortfall)}
\usage{
gpde.Vfun(par, dat, m)

gpde.phi(par, dat, V, m)

gpde.dphi(par, dat, V, m)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{y_m} and \eqn{\xi}, respectively the \eqn{m}-year return level and the shape parameter.}

\item{dat}{sample vector}

\item{m}{number of observations of interest for return levels. See \strong{Details}}

\item{V}{vector calculated by \code{gpde.Vfun}}
}
\description{
Vector implementing conditioning on approximate ancillary statistics for the TEM

Canonical parameter in the local exponential family approximation
}
\seealso{
\code{\link{gpde}}
}
\keyword{internal}

