% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpdN.infomat}
\alias{gpdN.infomat}
\title{Information matrix of the generalized Pareto distribution (mean of maximum of N exceedances parametrization)}
\usage{
gpdN.infomat(par, dat, N, method = c("obs", "exp"), nobs = length(dat))
}
\arguments{
\item{par}{vector of length 2 containing \eqn{z} and \eqn{\xi}, respectively the mean excess of the maxima of N exceedances above the threshold and the shape parameter.}

\item{dat}{sample vector}

\item{N}{block size for threshold exceedances.}
}
\description{
Information matrix of the generalized Pareto distribution (mean of maximum of N exceedances parametrization)
}
\seealso{
\code{\link{gpdN}}
}
\keyword{internal}
