% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpdN.mean}
\alias{gpdN.mean}
\title{This function returns the mean of N maxima from the GP.}
\usage{
gpdN.mean(par, N)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{z} and \eqn{\xi}, respectively the mean excess of the maxima of N exceedances above the threshold and the shape parameter.}

\item{N}{block size for threshold exceedances.}
}
\description{
This function returns the mean of N maxima from the GP.
}
\seealso{
\code{\link{gpdN}}
}
\keyword{internal}
