% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpdN.quant}
\alias{gpdN.quant}
\title{This function returns the qth percentile of N maxima from the GP.}
\usage{
gpdN.quant(par, q, N)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{z} and \eqn{\xi}, respectively the mean excess of the maxima of N exceedances above the threshold and the shape parameter.}

\item{N}{block size for threshold exceedances.}
}
\description{
This function returns the qth percentile of N maxima from the GP.
}
\seealso{
\code{\link{gpdN}}
}
\keyword{internal}
