% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpdN.temstat}
\alias{gpdN.temstat}
\alias{gpdN.Vfun}
\alias{gpdN.phi}
\alias{gpdN.dphi}
\title{Tangent exponential model statistics for the generalized Pareto distribution (mean of maximum of N exceedances parametrization)}
\usage{
gpdN.Vfun(par, dat, N)

gpdN.phi(par, dat, N, V)

gpdN.dphi(par, dat, N, V)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{z} and \eqn{\xi}, respectively the mean excess of the maxima of N exceedances above the threshold and the shape parameter.}

\item{dat}{sample vector}

\item{N}{block size for threshold exceedances.}

\item{V}{vector calculated by \code{gpdN.Vfun}}
}
\description{
Vector implementing conditioning on approximate ancillary statistics for the TEM

Canonical parameter in the local exponential family approximation
}
\seealso{
\code{\link{gpdN}}
}
\keyword{internal}
