% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tem.R
\name{gpdr.infomat}
\alias{gpdr.infomat}
\title{Observed information matrix for GP distribution (return levels)}
\usage{
gpdr.infomat(par, dat, m)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{y_m} and \eqn{\xi}, respectively the \eqn{m}-year return level and the shape parameter.}

\item{dat}{sample vector}

\item{m}{number of observations of interest for return levels. See \strong{Details}}
}
\description{
The information matrix is parametrized in terms of rate of \code{m}-year return level and shape
}
\seealso{
\code{\link{gpdr}}
}
\keyword{internal}

