% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpdr.ll}
\alias{gpdr.ll}
\alias{gpdr.ll.optim}
\title{Negative log likelihood of the generalized Pareto distribution (return levels)}
\usage{
gpdr.ll(par, dat, m)

gpdr.ll.optim(par, dat, m)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{y_m} and \eqn{\xi}, respectively the \eqn{m}-year return level and the shape parameter.}

\item{dat}{sample vector}

\item{m}{number of observations of interest for return levels. See \strong{Details}}
}
\description{
Negative log likelihood of the generalized Pareto distribution (return levels)

Negative log likelihood parametrized in terms of log return level and shape in order to perform unconstrained optimization
}
\seealso{
\code{\link{gpdr}}
}
\keyword{internal}
