% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpdr.score}
\alias{gpdr.score}
\title{Score of the profile log likelihood for the GP distribution (return levels parametrization)}
\usage{
gpdr.score(par, dat, m)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{y_m} and \eqn{\xi}, respectively the \eqn{m}-year return level and the shape parameter.}

\item{dat}{sample vector}

\item{m}{number of observations of interest for return levels. See \strong{Details}}
}
\description{
Score of the profile log likelihood for the GP distribution (return levels parametrization)
}
\seealso{
\code{\link{gpdr}}
}
\keyword{internal}
