% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgplikelihoods.R
\name{gpdtopar}
\alias{gpdtopar}
\title{Transformation from the generalized Pareto to unit Pareto}
\usage{
gpdtopar(dat, loc = 0, scale, shape, lambdau = 1)
}
\arguments{
\item{dat}{vector or matrix of data}

\item{loc}{vector of location parameters}

\item{scale}{vector of scale parameters, strictly positive}

\item{shape}{shape parameter}

\item{lambdau}{vector of probability of marginal threshold exceedance}
}
\value{
a vector or matrix of the same dimension as \code{dat} with unit Pareto observations
}
\description{
Transformation from the generalized Pareto to unit Pareto
}
\keyword{internal}
