% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgplikelihoods.R
\name{intensBR}
\alias{intensBR}
\title{Intensity function for the Brown-Resnick model}
\usage{
intensBR(tdat, Lambda, cholPrecis = NULL)
}
\arguments{
\item{tdat}{matrix of unit Pareto observations}

\item{Lambda}{conditionally negative definite parameter matrix of the Huesler--Reiss model}

\item{cholPrecis}{Cholesky root of the corresponding precision matrix \code{solve(Sigma)}. Default to \code{NULL}, meaning the latter is calculated within the function}
}
\value{
log intensity contribution
}
\description{
The intensity function includes the normalizing constants
}
\keyword{internal}
