% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgplikelihoods.R
\name{intensXstud}
\alias{intensXstud}
\title{Intensity function for the extremal Student model}
\usage{
intensXstud(tdat, df, Sigma, cholPrecis = NULL)
}
\arguments{
\item{tdat}{matrix of unit Pareto observations}

\item{df}{degrees of freedom, must be larger than 1}

\item{Sigma}{scale matrix}

\item{cholPrecis}{Cholesky root of the precision matrix \code{solve(Sigma)}. Default to \code{NULL}, meaning the latter is calculated within the function}
}
\value{
log intensity contribution
}
\description{
The intensity function includes the normalizing constants
}
\keyword{internal}
