% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taildep2.R
\name{kjtail}
\alias{kjtail}
\title{Estimators of the tail coefficient}
\usage{
kjtail(
  data,
  q,
  ptail = NULL,
  mqu,
  type = 1,
  ties.method = eval(formals(rank)$ties.method),
  ...
)
}
\arguments{
\item{data}{a matrix of observations}

\item{q}{vector of quantile levels}

\item{ptail}{tail probability smaller than \code{q}. Default to \code{NULL}}

\item{mqu}{marginal quantile levels for semiparametric estimation; data above this are modelled using a generalized Pareto distribution. If missing, empirical estimation is used throughout}

\item{type}{integer indicating the estimator type}

\item{ties.method}{method for ties}
}
\value{
a list with elements
\itemize{
\item \code{p} probability
\item \code{eta} estimated coefficient of tail dependence \eqn{\eta}
\item \code{eta_sd} estimated standard error of \eqn{\eta}
}
}
\description{
Estimators proposed by Krupskii and Joe under second order expansion
for the coefficient of tail dependence \eqn{\eta} and the
joint tail orthant probability
}
\examples{
rho <- runif(1, -1, 1)
d <- 2
data <- mev::mvrnorm(
   n = 1e5,
   mu = rep(0, d),
  Sigma = matrix(rho, d, d) + diag(1 - rho, d))
q <- seq(0.95, 0.995, by = 0.005)
taildep <- kjtail(data = data, q = q)
with(taildep,
 plot(x = 1-pat,
      y = eta,
      ylim = c(0,1),
      panel.first = {abline(h = (1+rho)/2)}))
}
