% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{leedspollution}
\alias{leedspollution}
\title{Leeds air pollution}
\format{
A data frame with 11455 rows and 8 variables:
\describe{
  \item{\code{date}}{[character] a date with format yyy-mm-dd}
  \item{\code{O3}}{[integer] ozone (in nanograms per cubic meter)}
  \item{\code{NO}}{[integer] nitrogen oxyde (in nanograms per cubic meter)}
  \item{\code{CO}}{[double] carbon monoxyde (in micrograms per cubic meter)}
  \item{\code{NO2}}{nitrogen dioxyde  (in nanograms per cubic meter)}
  \item{\code{SO2}}{sulphur dioxide  (in nanograms per cubic meter)}
  \item{\code{PM10}}{[integer] particulate matter 10, (in nanograms per cubic meter)}
  \item{\code{PM2.5}}{[integer] particulate matter 2.5, (in nanograms per cubic meter)}
}
}
\source{
Crown 2025 copyright Defra via \code{uk-air.defra.gov.uk}, licenced under the Open Government Licence (OGL).
}
\usage{
leedspollution
}
\description{
Daily maximum data (hourly for PM10) on air pollution for the Leeds Centre station in Yorkshire and Humberside station. The data goes from January 1st, 1993, until December 31st, 2024. Data show seasonality and there are some outliers. From December 2nd, 2008 onwards, particulate matters (PM10 and PM2.5) are measured using  a tapered element oscillating microbalance (TEOM) and Filter Dynamics Measurement System (FDMS). The data for PM2.5 is missing before the change of instrumentation. A total of 231 daily measurements with only missing values were removed during preprocessing.
}
\keyword{datasets}
