% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{lm.gev}
\alias{lm.gev}
\title{General linear models with generalized extreme value distribution}
\usage{
lm.gev(
  formula = list(loc = ~1, scale = ~1, shape = ~1),
  response,
  data = NULL,
  link = list(loc = identity, scale = log, shape = identity),
  invlink = list(loc = identity, scale = exp, shape = identity),
  start = NULL,
  scale = TRUE,
  method = c("nlminb", "BFGS"),
  show = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame with response and observations}

\item{start}{named list of starting values}

\item{method}{string indicating the outer optimization routine for the augmented Lagrangian. One of \code{nlminb} or \code{BFGS}.}
}
\value{
a list containing the following components:
\itemize{
\item \code{estimate} a vector containing the maximum likelihood estimates.
\item \code{std.err} a vector containing the standard errors.
\item \code{vcov} the variance covariance matrix, obtained as the numerical inverse of the observed information matrix.
\item \code{method} the method used to fit the parameter.
\item \code{nllh} the negative log-likelihood evaluated at the parameter \code{estimate}.
\item \code{convergence} components taken from the list returned by \code{\link[alabama]{auglag}}.
Values other than \code{0} indicate that the algorithm likely did not converge.
\item \code{counts} components taken from the list returned by \code{\link[alabama]{auglag}}.
\item \code{xdat} vector of data
}
}
\description{
This function returns an object of class \code{mev_gev}, with default methods for printing and quantile-quantile plots.
The default starting values are the solution of the probability weighted moments.
}
\examples{
data(venice, package = "mev")
proto <- lm.gev(
formula = list(
  loc = ~year, scale = ~1, shape = ~1),
  response = "r1",
  data = venice)
}
