% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nidd}
\alias{nidd}
\title{River Nidd Flow}
\format{
a vector of size 154
}
\source{
Natural Environment Research Council (1975). \emph{Flood Studies Report}, volume 4.  pp. 235--236.
}
\description{
The data consists of exceedances over the threshold 65 cubic meter per second of the River Nidd at Hunsingore Weir, for 35 years of data between 1934 and 1969.
}
\references{
Davison, A.C. and R.L. Smith (1990). Models for Exceedances over High Thresholds, \emph{Journal of the Royal Statistical Society. Series B (Methodological)}, \bold{52}(3), 393--442. With discussion.
}
\seealso{
\code{nidd.thresh} from the \code{evir} package
}
