% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{ordexp2gev}
\alias{ordexp2gev}
\title{Transformation of matrix from exponential to GEV order statistics
Given a matrix of exponential draws (ordered from largest to smallest),
apply a transformation to map them to the generalized extreme value distribution (GEV)
where now the rows have increasing order statistics from left to right}
\usage{
ordexp2gev(edat, par)
}
\arguments{
\item{edat}{matrix of exponential order statistics (ordered by row from largest to smallest)}

\item{par}{vector of location, scale and shape parameter of the GEV}
}
\value{
a matrix of the same dimension as \code{edat}
}
\description{
Transformation of matrix from exponential to GEV order statistics
Given a matrix of exponential draws (ordered from largest to smallest),
apply a transformation to map them to the generalized extreme value distribution (GEV)
where now the rows have increasing order statistics from left to right
}
\keyword{internal}
