% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pandemics}
\alias{pandemics}
\title{Deaths from pandemics}
\format{
A data frame with 72 rows and 8 variables:
\describe{
  \item{\code{event}}{name of the event}
  \item{\code{startyear}}{start year of the event}
  \item{\code{endyear}}{end year of the event}
  \item{\code{lower}}{lower bound on estimated deaths (in thousands)}
  \item{\code{average}}{average estimated deaths (in thousands)}
  \item{\code{upper}}{upper bound on estimated deaths (in thousands)}
  \item{\code{saverage}}{scaled average of estimated deaths (in thousands)}
  \item{\code{population}}{estimated population at risk (in thousands)}
}
}
\source{
Cirillo, P. and N.N. Taleb (2020). \emph{Tail risk of contagious diseases}. Nat. Phys. \bold{16}, 606–613 (2020). <doi:10.1038/s41567-020-0921-x>
}
\usage{
pandemics
}
\description{
The data base contains estimated records of the number of deaths from pandemics.
}
\keyword{datasets}
