% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{plot.extprof}
\alias{plot.extprof}
\title{Plot of (modified) profile likelihood}
\usage{
\method{plot}{extprof}(x, ...)
}
\arguments{
\item{x}{an object of class \code{extprof} returned by \code{\link{gpd.pll}} or \code{\link{gev.pll}}.}

\item{...}{further arguments to \code{plot}.}
}
\value{
a graph of the (modified) profile likelihoods
}
\description{
The function plots the (modified) profile likelihood and the tangent exponential profile likelihood
}
\references{
Brazzale, A. R., Davison, A. C. and Reid, N. (2007). \emph{Applied Asymptotics: Case Studies in Small-Sample Statistics}. Cambridge University Press, Cambridge.

Severini, T. A. (2000). \emph{Likelihood Methods in Statistics}. Oxford University Press, Oxford.
}
