% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{plot.mev_shape_rbm}
\alias{plot.mev_shape_rbm}
\title{Plots for random block maximum estimator}
\usage{
\method{plot}{mev_shape_rbm}(x, type = c("shape", "risk"), log = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{mev_shape_rbm} returned by \code{shape.rbm}}

\item{type}{[string] type of plot, either \code{"shape"} for the tail index or \code{"risk"} for the empirical Bayes risk}

\item{log}{[logical] if \code{TRUE} (default), the x-axis for the number of exceedances is displayed on the log scale.}

\item{...}{additional arguments, currently ignored}
}
\value{
one or more plots
}
\description{
The function returns plot of the shape estimator along with the value (and 95\% Wald-based confidence interval) at the selected threshold, or a plot of the empirical Bayes risk.
}
