% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vartymetric.R
\name{plot.mev_thselect_vmetric}
\alias{plot.mev_thselect_vmetric}
\title{Plots for Varty and al. metric-based threshold selection}
\usage{
\method{plot}{mev_thselect_vmetric}(
  x,
  type = c("qq", "pp", "exp", "metric"),
  B = 1000L,
  probs = c(0.025, 0.975),
  ...
)
}
\arguments{
\item{x}{an object of class \code{mev_thselect_vmetric} produced by a call to \code{thselect.vmetric}}

\item{type}{string; a single string indicating the choice of plot}

\item{B}{number of simulations for variability of estimation}

\item{probs}{quantile levels for intervals.}

\item{...}{additional arguments, currently ignored}
}
\value{
NULL; the function is used to produce a plot
}
\description{
This S3 method produces quantile-quantile plots with confidence and tolerance bands on various scale (uniform, exponential, generalized Pareto), or a plot of the metric as a function of the threshold.
}
