% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wdiag.R
\name{plot.mev_thselect_wadsworth}
\alias{plot.mev_thselect_wadsworth}
\title{Sequential analysis diagnostic plots for threshold selection}
\usage{
\method{plot}{mev_thselect_wadsworth}(x, type = c("wn", "ps"), ...)
}
\arguments{
\item{x}{object returned by a call to \code{thselect.wseq}}

\item{type}{string giving the plots to produce}

\item{...}{additional arguments passed to plotting function}
}
\value{
\code{NULL}; the method is used to generate plots
}
\description{
Sequential analysis diagnostic plots for threshold selection
}
