% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrl.R
\name{plot.mev_tstab_mrl}
\alias{plot.mev_tstab_mrl}
\title{Mean residual life parameter stability plot}
\usage{
\method{plot}{mev_tstab_mrl}(
  x,
  xlab = c("thresh", "nexc"),
  level = 0.95,
  type = c("band", "ptwise"),
  ...
)
}
\arguments{
\item{x}{object resulting from a call to \code{tstab.mrl}}

\item{xlab}{[string]; whether to plot mean residual life plot as a function of threshold value of number of exceedances}

\item{level}{[numeric] level of Wald confidence intervals}

\item{type}{[string] whether to plot pointwise confidence intervals using segments (\code{"ptwise"}) or using dashed lines (\code{"band"})}

\item{...}{additional arguments, currently ignored}
}
\value{
\code{NULL}; use to produce plots
}
\description{
Mean residual life parameter stability plot
}
