% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{power.vario}
\alias{power.vario}
\title{Power variogram model}
\usage{
power.vario(h, alpha, scale = 1)
}
\arguments{
\item{h}{vector or matrix of pairwise distances}

\item{alpha}{smoothness parameter}

\item{scale}{scale parameter}
}
\value{
a vector or matrix of variogram values of the same length as \code{h}
}
\description{
The power variogram model is
\deqn{\gamma(h) = (\|h\|/\lambda)^\alpha, \quad \lambda>0, \alpha \in [0,2).}
}
\keyword{internal}
