% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{powerexp.cor}
\alias{powerexp.cor}
\title{Power exponential correlation model}
\usage{
powerexp.cor(h, alpha = 1, scale = 1)
}
\arguments{
\item{h}{vector or matrix of pairwise distances}

\item{alpha}{smoothness parameter}

\item{scale}{scale parameter}
}
\value{
a vector or matrix of correlations of the same dimensions as \code{h}
}
\description{
The power correlation model is
\deqn{\rho(h) = \exp\{-(\|h\|/\lambda)^\alpha\}, \quad \lambda>0, \alpha \in [0,2).}
}
\keyword{internal}
