% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{pp.infomat}
\alias{pp.infomat}
\title{Information matrix for the Poisson process likelihood}
\usage{
pp.infomat(par, dat, method = c("obs", "exp"), u, np = 1,
  nobs = length(dat))
}
\arguments{
\item{par}{vector of \code{loc}, \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{method}{string indicating whether to use the expected  (\code{'exp'}) or the observed (\code{'obs'} - the default) information matrix.}

\item{u}{threshold}

\item{np}{number of periods of observations. This is a \emph{post hoc} adjustment for the intensity so that the parameters of the model coincide
with those of a generalized extreme value distribution with block size \code{length(dat)/np}.}

\item{nobs}{number of observations for the expected information matrix. Default to \code{length(dat)} if \code{dat} is provided.}
}
\description{
The function returns the expected or observed information matrix.
}
\note{
For the expected information matrix, the number of points above the threshold is random, but should correspond to
\code{np}\eqn{\Lambda}. In this sense, unless \code{np} is linear in the sample size, the
}
\seealso{
\code{\link{pp}}
}
\keyword{internal}
