% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condextr.R
\name{predict.condex}
\alias{predict.condex}
\title{Prediction from Heffernan-Tawn model}
\usage{
\method{predict}{condex}(fitted, B = 1e+05, nm = 10000)
}
\arguments{
\item{B}{number of Monte Carlo replications}

\item{alpha}{vector of \eqn{\alpha} parameters, recycled if necessary}

\item{beta}{vector of \eqn{\beta} parameters, recycled if necessary}

\item{data}{matrix of data}

\item{thresh}{vector of uniform thresholds}

\item{region}{risk region on uniform scale}

\item{constraint}{string indicating whether margins are supposed equiprobable
or the relative frequency with which a variable is largest is estimated empirically (multinomial)}

\item{type}{string; should}
}
\description{
Prediction from Heffernan-Tawn model
}
