% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condextr.R
\name{profile.condex}
\alias{profile.condex}
\title{Heffernan-Tawn model profile pseudo log likelihood}
\usage{
\method{profile}{condex}(
  fitted,
  which = c("alpha", "beta"),
  data,
  maxsteps = 101L,
  index = 1,
  thresh = 0.95
)
}
\arguments{
\item{alpha_grid}{grid of values for \eqn{\alpha}}

\item{beta_grid}{grid of values for \eqn{\beta}}
}
\description{
Heffernan-Tawn model profile pseudo log likelihood
}
