% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condsimu.R
\name{propRWMH}
\alias{propRWMH}
\title{Proposals for random walk Metropolis-Hastings}
\usage{
propRWMH(cur, cov, lbound = rep(-Inf, ncol(cov)), ubound = rep(Inf,
  ncol(cov)))
}
\description{
This function transforms a vector \code{cur} to an unconstrained scale based on \code{lbound} and \code{ubound},
then samples draws from a multivariate Gaussian vector with covariance matrix \code{cov} centered at the current (unconstrained) value.
The function then returns a list containing the proposal on the unconstrained scale (\code{trprop}),
on the original scale (\code{prop}) along with the log of the jacobian of the transformation (\code{logjac}).
}
