% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{qqplot.blocksize}
\alias{qqplot.blocksize}
\title{Diagnostic plots for max-stability based on blocks of GEV samples}
\usage{
qqplot.blocksize(
  xdat,
  type = c("max", "range", "all"),
  B = 1000L,
  marginal = FALSE,
  rounding = 0,
  lb = NULL,
  plot = TRUE,
  level = 0.95,
  np = NULL,
  simult = TRUE
)
}
\arguments{
\item{xdat}{\code{n} by \code{m} matrix of GEV observations, ordered by row from smallest to largest}

\item{type}{string; the statistic to return. Either the maximum of each row (\code{max}), the standardized difference between the penultimate and largest value (\code{spacing}), the ratio of maximum to spacing (\code{ratio}) or the whole sample (\code{all})}

\item{B}{number of bootstrap samples}

\item{marginal}{logical; if \code{TRUE}, estimates are based on the marginal likelihood of the \eqn{m-1} smallest order statistics of the sample}

\item{rounding}{amount of rounding}

\item{lb}{lower bound for left-censoring, default to \code{NULL} in absence}

\item{plot}{logical; if \code{TRUE} (default), returns a quantile-quantile plot}

\item{level}{confidence level for confidence and tolerance intervals}

\item{np}{number of points at which to evaluate quantile-quantile plots. Must be either \code{NULL}, or a vector of integer of the same length as \code{type} (otherwise it is recycled).}
}
\value{
a list with elements for building quantile-quantile plots, including
\itemize{
\item \code{plots} list of plots with elements \code{x}, \code{y}, a list \code{confint} with matrices \code{simultaneous} and \code{pointwise}, \code{type} of value and \code{distribution} (currently only uniform)
\item \code{mle}: maximum likelihood estimate of the location, scale, and shape
\item \code{param} \code{B} by 3 matrix of bootstrap parameter estimates
\item \code{type} vector of string with statistics
\item \code{bootstrap} type of bootstrap, only \code{parametric} for now
\item \code{n} number of rows of \code{xdat}
\item \code{m} number of columns of \code{xdat} for comparison
\item \code{marginal} logical; if \code{TRUE}, uses the marginal likelihood of the \eqn{m-1} smallest order statistics per block for estimation
\item \code{icens} logical; if \code{TRUE}, data treated as rounded (interval-censored)
\item \code{lcens} logical; if \code{TRUE}, data are left-censored below \code{lb}
\item \code{lb} lower bound for left-censoring
\item \code{rounding} double \eqn{\delta} indicating the amount of rounding, assuming \eqn{\delta/2} on either size of the reported value
\item \code{xdat} matrix of original observations
}
}
\description{
Given a sample of ordered GEV draws, calculate the ingredients of diagnostic
quantile-quantile plots using the bootstrap
}
\examples{
xdat <- build.blocks(mev::rgev(n = 50), m = 2)
\dontrun{
qqplot.blocksize(xdat, type = "max", marginal = TRUE, B = 100)
}
}
