% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{qqplot.unif}
\alias{qqplot.unif}
\title{Pointwise and simultaneous binomial confidence intervals for uniform via simulation}
\usage{
qqplot.unif(xdat, K = 100, B = 1000, level = 0.95, plot = TRUE)
}
\arguments{
\item{xdat}{vector of \code{N} postulated uniform samples, obtained by applying the ECDF}

\item{K}{number of evaluation points for the plotting positions}

\item{B}{number of Monte Carlo samples}

\item{level}{vector of pointwise and simultaneous confidence levels, recycled if necessary}

\item{plot}{logical; if \code{TRUE}, produce a plot of the empirical distribution function}
}
\description{
Given a vector of draws transformed using the probability integral transform scale
to what should be uniform positions, produce plots with pointwise and simultaneous confidence intervals for uniformity.
}
\examples{
xdat <- runif(200)
qqplot.unif(xdat)
}
\references{
Sailynoja, T., Burkner, P.C. and Vehtari, A. (2022). Graphical test for discrete uniformity and its applications in goodness-of-fit evaluation and multiple sample comparison, \emph{Statistics and Computing}, 32, \doi{10.1007/s11222-022-10090-6}
}
