% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condsimu.R
\name{rcondmvtnorm}
\alias{rcondmvtnorm}
\title{Conditional distribution of Gaussian or Student subvectors}
\usage{
rcondmvtnorm(n = 1, ind, x, lbound = rep(-Inf, length(ind)),
  ubound = rep(Inf, length(ind)), mu, Sigma, df = NULL,
  model = c("norm", "stud"))
}
\arguments{
\item{n}{sample size. Default to 1.}

\item{ind}{a \code{d} vector of indices to impute with integer entries in \eqn{\{1, \ldots, d+p\}}}

\item{x}{a \code{p} vector with the values of process at remaining coordinates}

\item{lbound}{\code{d} vector of lower bounds}

\item{ubound}{\code{d} vector of upper bounds for truncated}

\item{mu}{\code{d+p} vector of location parameters}

\item{Sigma}{\code{d+p} by \code{d+p} scale matrix}

\item{df}{degrees of freedom of the \code{d+p} dimensional Student process}

\item{model}{string indicating family, either \code{norm} for Gaussian or \code{stud} for Student-t}
}
\value{
an n by d matrix of conditional simulations
}
\description{
The function samples (truncated) Gaussian or Student vectors
corresponding to indices \code{ind} given the values at the remaining index.
The location vector \code{mu} and the scale matrix \code{Sigma} are those of the \eqn{d+p} vector.
The routine relies on the CDF approximation based on minimax exponential tilting implemented in the \code{TruncatedNormal} package.
}
