% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condextr.R
\name{residuals.condex}
\alias{residuals.condex}
\title{Residuals from conditional extremes model}
\usage{
\method{residuals}{condex}(fitted, data, thresh)
}
\arguments{
\item{data}{matrix of observations}

\item{thresh}{vector of thresholds on the uniform scale}

\item{alpha}{scale parameter}

\item{beta}{power parameter}

\item{testIndep}{logical; if \code{TRUE}, compute a test of independence using energy statistics for each conditioning variable in turn}

\item{group}{integer vector for the group; only observations in group 1 are considered considered to build residuals.}
}
\description{
Residuals from conditional extremes model
}
