% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{rho.ghp}
\alias{rho.ghp}
\title{Second order tail index estimator of Gomes et al.}
\usage{
rho.ghp(xdat, k, alpha = 2)
}
\arguments{
\item{xdat}{vector of positive observations}

\item{k}{number of highest order statistics to use for estimation}

\item{alpha}{positive scalar tuning parameter}
}
\description{
Estimator of the second order regular variation parameter \eqn{\rho \leq 0} parameter for heavy-tailed data proposed by Gomes et al. (2003)
}
\references{
Gomes, M.I., de Haan, L. & Peng, L. (2002). Semi-parametric Estimation of the Second Order Parameter in Statistics of Extremes. \emph{Extremes} 5, 387–414. <doi:10.1023/A:1025128326588>
}
