% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{rlarg.score}
\alias{rlarg.score}
\title{Score of the r-largest observations}
\usage{
rlarg.score(par, dat)
}
\arguments{
\item{par}{vector of \code{loc}, \code{scale} and \code{shape}}

\item{dat}{an \code{n} by \code{r} sample matrix, ordered from largest to smallest in each row}
}
\value{
score vector of size 3
}
\description{
The score is computed via linear interpolation for the shape parameter in a neighborhood of zero
}
\seealso{
\code{\link{rlarg}}
}
\keyword{internal}
