% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmar.R
\name{rmar1}
\alias{rmar1}
\title{Simulation from first-order max-autoregressive processes}
\usage{
rmar1(n, theta, shape = 0)
}
\arguments{
\item{n}{sample size}

\item{theta}{extremal index, a value in (0,1]}

\item{shape}{non-negative shape parameter of the GEV}
}
\value{
a vector of length \code{n} drawn from the stationary distribution.
}
\description{
Generate data from stationary sequences for extremes for
non-negative shapes, following Tavares (1977) for the Gumbel case.
}
\details{
The models are parametrized in terms of extremal index \eqn{\theta \in (0,1]}.

When \code{shape = 0}, the stationary process has unit Gumbel margins.
When \code{shape > 0}, the margins have Frechet margins with distribution \eqn{F(x) = \exp(-x^{-1/\xi}).}
}
\examples{
X1 <- rmar1(n = 1000, theta = 0.5)
X2 <- rmar1(n = 1000, theta = 0.2, shape = 0.2)
par(mfrow = c(1, 2))
plot(X1)
plot(X2)
xacf(X1, qlev = 0.9)
xacf(X2, qlev = 0.9)
}
\references{
Valadares Tavares, L. (1977). The Exact Distribution of Extremes of a Non-Gaussian Process. \emph{Stochastic Processes and Their Applications} (2): 151-56. \doi{10.1016/0304-4149(77)90026-6}

Davis, Richard A., and Sidney I. Resnick (1989). Basic Properties and Prediction of Max-ARMA Processes, \emph{Advances in Applied Probability}, 21 (\bold{4}): 781–803. \doi{10.2307/1427767}.
}
