% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmnorm}
\alias{rmnorm}
\title{Multivariate Normal distribution sampler}
\usage{
rmnorm(n, mu, Sigma)
}
\arguments{
\item{n}{sample size}

\item{mu}{mean vector. Will set the dimension}

\item{Sigma}{a square covariance matrix, of same dimension as \code{mu}.
No sanity check is performed to validate that the matrix is positive definite, so use at own risk}
}
\value{
an \code{n} sample from a multivariate Normal distribution
}
\description{
Sampler derived using the eigendecomposition of the covariance
matrix \code{Sigma}. The function uses the Armadillo random normal generator
}
\examples{
rmnorm(n = 10, mu = c(0,2), Sigma = diag(2))
}
