% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{rrlarg}
\alias{rrlarg}
\title{Simulate r-largest observations from point process of extremes}
\usage{
rrlarg(n, r, loc, scale, shape)
}
\arguments{
\item{n}{sample size}

\item{r}{number of observations per block}

\item{loc}{location parameter}

\item{scale}{scale parameter}

\item{shape}{shape parameter}
}
\value{
an \code{n} by \code{r} matrix of samples from the point process, ordered from largest to smallest in each row.
}
\description{
Simulate the \code{r}-largest observations from a Poisson point process with intensity
\deqn{\Lambda(x) = (1+\xi(x-\mu)/\sigma)^{-1/\xi}}.
}
