% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{shape.genjack}
\alias{shape.genjack}
\title{Generalized jackknife shape estimator}
\usage{
shape.genjack(xdat, k)
}
\arguments{
\item{xdat}{vector of positive observations}

\item{k}{vector of order statistics; if missing, a vector going from 10 to sample size minus one.}
}
\value{
a data frame with the number of order statistics \code{k} and the shape parameter estimate \code{shape}, or a single numeric value if \code{k} is a scalar integer.
}
\description{
Generalized jackknife shape estimator
}
\references{
Gomes, I.M., João Martins, M. and Neves, M.  (2000) Alternatives to a Semi-Parametric Estimator of Parameters of Rare Events-The Jackknife Methodology. \emph{Extremes}, 3, 207–229. \doi{10.1023/A:1011470010228}
}
