% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{shape.genpickands}
\alias{shape.genpickands}
\title{Generalized Pickands estimator}
\usage{
shape.genpickands(xdat, k, cst = 0.75, constrained = FALSE, ...)
}
\arguments{
\item{xdat}{vector of observations}

\item{k}{vector of the number of exceedances for the threshold}

\item{cst}{double in the unit interval, the constant for the fraction. Default to 0.75}

\item{constrained}{logical; if \code{TRUE}, implements the constrained estimator taking the second-order parameters.}
}
\value{
a data frame with the number of order statistics \code{k} and the shape parameter estimate \code{shape}, or a single numeric value if \code{k} is a scalar.
}
\description{
Implements the generalized Pickands shape parameter estimator with or without constraints. The estimator obtains an initial value for the shape by taking \eqn{\xi=0} and \eqn{\rho=-1}.
}
\examples{
par(mfrow = c(1,2))
shape0 <- runif(1,-0.5,0.5)
xdat <- mev::rgp(n = 500, shape = shape0)
est_unc <- shape.genpickands(
 xdat, k = 15:400,
 constrained = FALSE)
 plot(
  est_unc,
  type = "l",
  panel.first = {abline(h = shape0)},
  ylim = c(-0.75,0.75)
  )
est_con <- shape.genpickands(
 xdat, k = 15:400,
 constrained = TRUE)
 plot(
  est_con,
  type = "l",
  panel.first = {abline(h = shape0)},
  ylim = c(-0.75,0.75)
  )
}
\references{
Johan Segers (2005). Generalized Pickands estimators for the extreme value index, Journal of Statistical Planning and Inference, 128(2), 381-396, <doi:10.1016/j.jspi.2003.11.004>
}
