% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lthill.R
\name{shape.lthill}
\alias{shape.lthill}
\title{Left-trimmed Hill estimator for the shape parameter}
\usage{
shape.lthill(xdat, k, k0 = k, vectorize = FALSE, sorted = FALSE)
}
\arguments{
\item{xdat}{[numeric] vector of positive observations}

\item{k}{[integer] number of order statistics for the threshold}

\item{k0}{[integer] scalar number of largest order statistics, no greater than \code{k}}

\item{vectorize}{[logical] if \code{TRUE}, computes for each value of}

\item{sorted}{[logical] if \code{TRUE}, data are assumed to be sorted in decreasing order}
}
\value{
a scalar with the shape parameter estimate
}
\description{
Given a sample of Pareto-tailed samples (positive tail index),
compute the left-trimmed Hill estimator. If \eqn{k0=k}, the estimator
reduces to Hill's estimator for the shape index
}
\examples{
# Pareto sample
n <- 200
xdat <- 10/(1 - runif(n)) - 10
shape.lthill(xdat = xdat, k = 100, k0 = 1:100)
}
