% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{shape.moment}
\alias{shape.moment}
\title{Dekkers and de Haan moment estimator for the shape}
\usage{
shape.moment(xdat, k)
}
\arguments{
\item{xdat}{vector of positive observations of length \eqn{n}}

\item{k}{number of largest order statistics}
}
\value{
a data frame with the number of order statistics \code{k} and the shape parameter estimate \code{shape}, or a single numeric value if \code{k} is a scalar.
}
\description{
Given a sample of exceedances, compute the moment estimator of the real shape parameter.
}
\references{
Dekkers, A.L.M. and de Haan, L. (1989). \emph{On the estimation of the extreme-value index and large quantile estimation.}, Annals of Statistics, \bold{17}, 1795-1833.
}
