% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{shape.pickands}
\alias{shape.pickands}
\title{Pickand's shape estimator}
\usage{
shape.pickands(xdat, k)
}
\arguments{
\item{xdat}{vector of positive observations of length \eqn{n}}

\item{k}{number of largest order statistics}
}
\value{
a data frame with the number of order statistics \code{k} and the shape parameter estimate \code{shape}, or a single numeric value if \code{k} is a scalar.
}
\description{
Given a sample of size \code{n} of positive exceedances, compute
the real shape parameter \eqn{\xi} based on the \code{k} largest order statistics.
}
\references{
Pickands, III, J. (1975). \emph{Statistical inference using extreme order statistics.} Annals of Statistics, \bold{3}, 119-131.
}
