% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{shape.vries}
\alias{shape.vries}
\title{de Vries shape estimator}
\usage{
shape.vries(xdat, k)
}
\arguments{
\item{xdat}{vector of positive observations of length \eqn{n}}

\item{k}{number of largest order statistics}
}
\value{
a data frame with the number of order statistics \code{k} and the shape parameter estimate \code{shape}, or a single numeric value if \code{k} is a scalar.
}
\description{
Given a sample of exceedances, compute the moment estimator of the positive shape parameter using the ratio of log ratio of exceedance and it's square.
}
\references{
de Haan, L. and Peng, L. (1998). \emph{Comparison of tail index estimators}, Statistica Neerlandica 52, 60-70.

de Haan, L. and Peng, L. (1998). Comparison of tail index estimators. \emph{Statistica Neerlandica}, 52: 60-70. \doi{10.1111/1467-9574.00068}
}
