% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stein.R
\name{stein_gp_lik}
\alias{stein_gp_lik}
\title{Stein's weighted generalized Pareto likelihood}
\usage{
stein_gp_lik(pars, xdat, weights = rep(1, length(xdat)), ...)
}
\arguments{
\item{pars}{vector of length with scale and shape parameters}

\item{xdat}{vector of exceedances in decreasing order if \code{sort=FALSE}}

\item{weights}{vector of weights}

\item{...}{additional arguments, currently ignored}
}
\description{
Stein's weighted generalized Pareto likelihood
}
\references{
Stein, M.L. (2023). A weighted composite log-likelihood approach to parametric estimation of the extreme quantiles of a distribution. \emph{Extremes} \bold{26}, 469-507 <doi:10.1007/s10687-023-00466-w>
}
\keyword{internal}
