% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stein.R
\name{Stein_weights}
\alias{Stein_weights}
\title{Stein's vector of weights}
\usage{
Stein_weights(n, gamma = 1)
}
\arguments{
\item{n}{integer, sample size}

\item{gamma}{positive scalar for power}
}
\value{
a vector of positive weights of length \code{n}
}
\description{
Computes a vector of decreasing weights for exceedances.
}
\references{
Stein, M.L. (2023). A weighted composite log-likelihood approach to parametric estimation of the extreme quantiles of a distribution. \emph{Extremes} \bold{26}, 469-507 <doi:10.1007/s10687-023-00466-w>
}
\keyword{internal}
