% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{test.blocksize}
\alias{test.blocksize}
\title{Likelihood ratio test for max-stability}
\usage{
test.blocksize(xdat, rounding = 0, alternative = c(1L, 2L, 3L), lb = NULL)
}
\arguments{
\item{xdat}{\code{n} by \code{m} matrix of observations assumed to arise from a GEV, ordered by row}

\item{rounding}{double, a positive number indicating the amount of censoring (e.g., 0.1 or 1)}

\item{alternative}{integer; \code{1} for a single shape parameter with parameters obtained from max-stability, \code{2} for a common shape parameter, but with free location and scale.}

\item{lb}{lower bound for left-censoring; default to none (\code{NULL})}
}
\value{
a data frame containing likelihood ratio statistics (\code{stat}), the degrees of freedom, a vector of p-values \code{pval} and the name of the \code{alternative}.
}
\description{
Given a matrix of block maxima split into blocks of size \code{m},
calculate test statistics and return p-values based on the asymptotic
chi-square distribution.
}
\examples{
samp <- build.blocks(mev::rgev(50, scale = 10), m = 4)
# x=-5 is approximately the 0.2 quantile of the above
test.blocksize(xdat = round(samp, 0), rounding = 1, lb = -5)
test.blocksize(xdat = round(samp, 0), rounding = 1)
test.blocksize(xdat = samp)
test.blocksize(xdat = samp, lb = -5, alternative = 1L)
}
