% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocktest.R
\name{test.unif}
\alias{test.unif}
\title{Durbin's mapping of uniforms}
\usage{
test.unif(xdat)
}
\arguments{
\item{xdat}{sample of uniform variates}
}
\value{
a data frame with elements
\itemize{
\item \code{test} string indicating which of the modified probability product, median and (one-sided) Kolmogorov--Smirnov tests
\item \code{stat} the test statistics
\item \code{pval} for the p-values from the null distributions
}
}
\description{
Given a set of uniform variates, obtained for example through application
of the probability integral  transform, consider alternative scaling which
are also uniform but may lead to more powerful tests.
}
\examples{
test.unif(runif(1000))
test.unif(rbeta(1000, 1, 1.5))
}
\references{
Durbin, J. (1961). Some Methods of Constructing Exact Tests. \emph{Biometrika}, 48(\bold{1/2}), 41–55. \doi{10.2307/2333128}
}
\keyword{internal}
