% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{thames}
\alias{thames}
\title{Thames river flow at Kingston}
\format{
A data frame with 142 rows and 3 variables:
\describe{
  \item{\code{date}}{date of measurement}
  \item{\code{flow}}{double maximum daily river flow (in cubic meter per seconds)}
  \item{\code{flag}}{logical; if \code{TRUE}, the value represents the instantaneous annual maximum, otherwise the natural annual max mean daily flow}
}
}
\source{
Acknowledgement: Data from the UK National River Flow Archive, \url{https://nrfa.ceh.ac.uk/data/station/info/39001}, extracted March 2026
}
\usage{
thames
}
\description{
Time series of annual maximum daily peak flow (in meter per seconds) of the Thames River at Kingston.
The 1894 record was modified as the previous value reported of 1064 cubic meter per second was
considered to be an overestimate of the true flow.
}
\keyword{datasets}
