% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.diag.R
\name{thselect.dobre}
\alias{thselect.dobre}
\title{Robust threshold selection method via OBRE}
\usage{
thselect.dobre(xdat, thresh, nsim = 99L)
}
\arguments{
\item{xdat}{vector of observations}

\item{thresh}{vector of candidate thresholds}

\item{nsim}{number of bootstrap simulations}
}
\value{
a list with elements
\itemize{
\item \code{coef} matrix of parameter estimates for the scale and shape parameters, estimated via OBRE
\item \code{thresh} vector of ordered thresholds
\item \code{pval} p-values from the procedure
\item \code{ndw} number of downweighted observations
}
}
\description{
The function calculates the cumulative sum of the non-zero weights for the generalized Pareto model fitted via optimal B-robust estimation (OBRE). Bootstrap samples from the generalized Pareto are generated and the sum of weights for the same numbers are calculated. This exploits the fact that the weights are monotonically increasing, so downweighting (if any) occurs in the tails with increasing weights (capped at unity).
}
\references{
Dupuis, D.J. (1998). Exceedances over High Thresholds: A Guide to Threshold Selection,
\emph{Extremes}, \bold{1}(3), 251--261.
}
