% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_exponential_thselect.R
\name{thselect.gbw}
\alias{thselect.gbw}
\title{Kernel-based threshold selection of Goegebeur, Beirlant and de Wet (2008)}
\usage{
thselect.gbw(
  xdat,
  kmax,
  kernel = c("Jackson", "Lewis"),
  rho = c("gbw", "ghp", "fagh", "dk"),
  ...
)
}
\arguments{
\item{xdat}{[vector] sample exceedances}

\item{kmax}{[int] maximum number of exceedances considered}

\item{kernel}{[string] kernel choice, one of \code{Jackson} or \code{Lewis}}

\item{rho}{string for the estimator of the second order regular variation. Can also be a negative scalar}
}
\value{

}
\description{
Kernel-based threshold selection of Goegebeur, Beirlant and de Wet (2008)
}
\references{
Goegebeur , Y., Beirlant , J., and de Wet , T. (2008). Linking Pareto-Tail Kernel Goodness-of-fit Statistics with Tail Index at Optimal Threshold and Second Order Estimation. REVSTAT-Statistical Journal, 6(\bold{1}), 51–69. <doi:10.57805/revstat.v6i1.57>
}
