% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickands.R
\name{thselect.mahadist}
\alias{thselect.mahadist}
\title{Mahalanobis distance-based methodology}
\usage{
thselect.mahadist(xdat, thresh, approx = c("asymptotic", "mc"), B = 1000L)
}
\arguments{
\item{approx}{[string] method to use to obtain moments of first L-moment}

\item{B}{[integer] number of replications for Monte Carlo approximation}
}
\value{
a vector of length 2 with the selected threshold and a corresponding \emph{p}-value based on a chi-square approximation to the test statistic
}
\description{
Compute the Mahalanobis distance-based threshold method over a grid of thresholds by
transforming data from generalized Pareto to unit exponential based on probability weighted moment estimates,
then computing the first L-moment and the L-skewness. The latter are compared to the
theoretical counterparts from a unit exponential sample of the same size, which is used to compute
the Mahalanobis distance. The threshold returned is the one which minimizes the distance.
}
\references{
Kiran, K. G. and Srivinas, V.V. (2021). \emph{A Mahalanobis distance-based automatic threshold selection method for peaks over threshold model.} Water Resources Research 57. <doi:10.1029/2020WR027534>
}
