% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thselect.R
\name{thselect.mdps}
\alias{thselect.mdps}
\title{Minimum distance threshold selection procedure}
\usage{
thselect.mdps(xdat)
}
\arguments{
\item{xdat}{vector of positive exceedances}
}
\value{
a list with components
\describe{
\item{\code{k0}}{order statistic corresponding to threshold (number of exceedances)}
\item{\code{shape}}{Hill's estimator of the tail index based on k0 exceedances}
\item{\code{thresh0}}{numerical value of the threshold, the n-k0+1 order statistic of the original sample}
}
}
\description{
Minimum distance threshold selection procedure
}
\references{
Clauset, A., Shalizi, C.R. and Newman, M.E.J. (2009). \emph{Power-Law Distributions in Empirical Data}. SIAM Review. Society for Industrial and Applied Mathematics, \bold{51}, 661-703, \doi{10.1137/070710111}
}
