% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickands.R
\name{thselect.pickands}
\alias{thselect.pickands}
\title{Pickands' order statistics threshold selection method}
\usage{
thselect.pickands(xdat, method = c("mle", "lmom", "quartiles"))
}
\arguments{
\item{xdat}{[numeric] vector of observations}

\item{method}{[string] estimation method, either the quartiles of Pickands (1975), maximum likelihood, probability weighted moments or L-moments}
}
\value{
a vector with elements \code{k}, the number of order statistics to keep, and \code{thresh}, the numerical value of the latter
}
\description{
Restricting to the largest fourth of the data, returns the number of exceedances that minimizes the Kolmogorov-Smirnov statistic, i.e., the maximum absolute difference between the estimated generalized Pareto and the empirical distribution of exceedances. Relative to the paper, different estimation methods are proposed.
}
\note{
The quartile esitmator of Pickands is robust, but very inefficient. It is provided for historical reasons.
}
\references{
James Pickands III (1975). \emph{Statistical inference using extreme order statistics}, Annals of Statistics, 3(\bold{1}) 119-131, \url{https://doi.org/10.1214/aos/1176343003}
}
