% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{thselect.rbm}
\alias{thselect.rbm}
\title{Threshold selection for the random block maxima method}
\usage{
thselect.rbm(xdat, kmax = length(xdat))
}
\arguments{
\item{xdat}{[vector] sample exceedances}

\item{kmax}{maximum number of exceedances to consider.}
}
\value{
a list with elements
\itemize{
\item{\code{k0}:} the number of exceedances at the selected threshold
\item{\code{thresh0}:} the selected threshold, or accordingly the (\eqn{k_0+1})th order statistic
\item{\code{shape}:} the RBM shape estimate
}
}
\description{
Threshold selection for the random block maxima method
}
