% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thselect.R
\name{thselect.samsee}
\alias{thselect.samsee}
\title{Threshold selection via SAMSEE}
\usage{
thselect.samsee(xdat)
}
\arguments{
\item{xdat}{vector of positive exceedances}
}
\value{
a list with elements
\describe{
\item{\code{k0}}{optimal number of exceedances}
\item{\code{shape}}{Hill estimator of the tail index}
\item{\code{thresh0}}{selected threshold}
}
}
\description{
Smooth asymptotic mean squared error estimator
of Schneider et al. (2021) for threshold selection.
The implementation uses a second-order regular variation index of \eqn{-1}.
}
\references{
Schneider, L.F., Krajina, A. and Krivobokova, T. (2021). \emph{Threshold selection in univariate extreme value analysis}, Extremes, \bold{24}, 881-913 \doi{10.1007/s10687-021-00405-7}
}
