% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thselect.R
\name{thselect.sihs}
\alias{thselect.sihs}
\title{Smooth inverse Hill threshold selection}
\usage{
thselect.sihs(xdat, kmax = 500L, method = "gamm", plot = FALSE, ...)
}
\arguments{
\item{xdat}{vector of positive exceedances}

\item{kmax}{maximum number of exceedances for the method, default to 500.}

\item{method}{method for smoothing: currently only GAMM with AR(1) correlation is supported}

\item{plot}{logical; if \code{TRUE}, returns a plot of the inverse Hill statistic}
}
\value{
a list with components
\describe{
\item{\code{k0}}{order statistic corresponding to threshold (number of exceedances)}
\item{\code{shape}}{Hill's estimator of the tail index based on k0 exceedances}
\item{\code{thresh0}}{numerical value of the threshold, the \eqn{n-k_0+1} order statistic of the original sample}
}
}
\description{
Threshold selection via smooth inverse Hill statistic. We use a simple generalized additive model with first-order autoregressive structure via package \code{mgcv}. As the latter is computationally intensive, it is recommended to limit the number of order statistics considered.
}
\references{
Schneider, L.F., Krajina, A. & Krivobokova, T. (2021). \emph{Threshold selection in univariate extreme value analysis}, Extremes, \bold{24}, 881–913 \url{https://doi.org/10.1007/s10687-021-00405-7}
}
