% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thselect.R
\name{thselect.wcvm}
\alias{thselect.wcvm}
\title{Threshold selection via minimization of the weighted Cramér-von Mises}
\usage{
thselect.wcvm(xdat, k)
}
\description{
For a Pareto-type sample, return the threshold that
minimizes a weighted Cramér-von Mises criterion for the
exponential sample with scale \eqn{H_{n, n_u}} and
the log increments.
}
\references{
Goegebeur , Y., Beirlant , J., and de Wet , T. (2008). Linking Pareto-Tail Kernel Goodness-of-fit Statistics with Tail Index at Optimal Threshold and Second Order Estimation. REVSTAT-Statistical Journal, 6(\bold{1}), 51–69. <doi:10.57805/revstat.v6i1.57>
}
