% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{w1500m}
\alias{w1500m}
\title{Best 200 times of Women 1500m Track}
\format{
a vector of size 200
}
\source{
<http://www.alltime-athletics.com/w_1500ok.htm>, accessed 14.08.2018
}
\description{
200 all-time best performance (in seconds) of women 1500-meter run.
}
