% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremogram.R
\name{xacf}
\alias{xacf}
\title{Extremogram}
\usage{
xacf(
  x,
  qlev,
  lag.max = NULL,
  plot = TRUE,
  confint = FALSE,
  B = 100L,
  level = 0.95,
  ties.method = "random",
  na.action = na.fail
)
}
\arguments{
\item{x}{vector of observations or time series}

\item{qlev}{quantile level of threshold, a scalar between (0,1)}

\item{lag.max}{integer, maximum lag at which to calculate the extremogram. Default to \eqn{10\log_{10}(n)}}

\item{plot}{logical; if \code{TRUE}, return a plot of the extremogram}

\item{confint}{logical; if \code{TRUE}, calculate \code{level} pointwise confidence intervals under independence, using a permutation-based approach}

\item{B}{integer, number of simulations for \code{confint}}

\item{level}{confidence level requested (default to 0.95).}

\item{ties.method}{string indicating the type of method for rank; see \code{\link[base]{rank}} for a list of options. Default to \code{"random"}}

\item{na.action}{function to be called to handle missing values}
}
\value{
a list with elements \code{extremogram} for the estimate of tail correlation at different lags, \code{upper} for the upper bound of the confidence interval for independent data and \code{level} of the latter.
}
\description{
Given a regular time series of observations,
compute the pairwise tail correlation between series
at different lags.
Permutation-based resampling are used to construct
confidence envelope (one-sided) for comparison with the independent
setting if \code{confint = TRUE}.
}
\examples{
xacf(x = rmar1(n = 1000, theta = 0.2, shape = 0.5),
     qlev = 0.95)
}
\references{
Davis, R. A., Mikosch, T., and Cribben, I. (2012). Towards estimating extremal serial dependence via the bootstrapped extremogram. Journal of Econometrics, 170(1), 142-152, \doi{10.1016/j.jeconom.2012.04.003}.

Davis, R. A. and T. Mikosch (2009). The extremogram: A correlogram for extreme events, \emph{Bernoulli}, 15(\bold{4}), 977-1009, \doi{10.3150/09-BEJ213}.
}
