% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickands.R
\name{xdep.pickands}
\alias{xdep.pickands}
\title{Estimator of the Pickands dependence function}
\usage{
xdep.pickands(xdat, w, qlev, region = c("sum", "max", "min"))
}
\arguments{
\item{xdat}{\code{n} by \code{d} matrix of observations}

\item{w}{\code{m} by \code{d} matrix of angles on the unit simplex}

\item{qlev}{quantile level for the threshold}

\item{region}{risk region for determining angles}
}
\value{
a vector of length \code{m} with Pickands dependence function
}
\description{
This function computes the nonparametric angular measure
of the multivariate observations using a empirical rank
transformation to unit Frechet margins, and from there
calculates the weights for the self-concordant empirical
likelihood imposing the mean constraint on the angular measure
}
\details{
\deqn{ \hat{A}(\boldsymbol{t}) = \sum_{i=1}^n p_i \max_{j=1}^d (w_{i,j}i t_j), \quad t \in \mathbb{S}_d}
}
\examples{
ang <- seq(0,1, by = 0.01)
xdat <- rmev(n = 1000, d = 2, param = 0.4)
pickands <- xdep.pickands(xdat, w = ang, qlev = 0)
plot(type = "n", x = 0.5,
     y = 1,
     xlim = c(0, 1),
     ylim = c(0.5, 1),
     xlab = "w",
     ylab = "Pickands dependence measure",
     bty = "l")
segments(x1 = 0.5, x0 = 0, y1 = 0.5, y0 = 1, col = "gray90")
segments(x1 = 0.5, x0 = 1, y1 = 0.5, y0 = 1, col = "gray90")
segments(x1 = 0, x0 = 1, y1 = 1, y0 = 1, col = "gray90")
lines(ang, pickands, lwd = 2)
}
