% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpdm.R
\name{xdep.tpdm}
\alias{xdep.tpdm}
\title{Tail pairwise dependence matrix}
\usage{
xdep.tpdm(
  xdat,
  qlev,
  ties.method = "random",
  margtrans = c("emp", "none"),
  standardize = TRUE
)
}
\arguments{
\item{xdat}{matrix of observations}

\item{qlev}{quantile level of threshold for the radial component}

\item{ties.method}{method for ties; see \link[base]{rank} for more details}

\item{margtrans}{string; if \code{"emp"} (default), apply a rank transformation to map observations to unit Frechet margins}

\item{standardize}{logical; if \code{TRUE} (default), matrix is standardized to correlation matrix}
}
\value{
a positive definite matrix
}
\description{
Given a multivariate sample of observations and a radial threshold quantile level,
estimate the tail pairwise dependence matrix (TPDM) empirically through marginal
transformation of the margins to unit Fréchet scale. This is entirely equivalent to
taking the definition of Cooley and Thibaud (2019) with GEV(1,0.5,0.5) margins.
}
\examples{
d <- 4L
xdat <- rmev(n = 1000, d = d, param = 0.9)
xdep.tpdm(xdat = xdat, qlev = 0.5, margtrans = "none")
# Equicorrelation matrix
Sigma <- 0.9 * diag(d) + matrix(0.1, d, d)
xdat <- rmnorm(n = 10000, mu = rep(0, d), Sigma = Sigma)
xdep.tpdm(xdat = xdat, qlev = 0.99)
}
\references{
Larsson, M. and S.I. Resnick(2012). \emph{Extremal dependence measure and extremogram: the regularly varying case}, Extremes 15, 231–256. <doi:10.1007/s10687-011-0135-9>

Cooley, D. and E. Thibaud (2019). \emph{Decompositions of dependence for high-dimensional extremes}, Biometrika, 106(\bold{3}), 587-604. <doi:10.1093/biomet/asz028>

Kiriliouk, A. and C. Zhou (2024+) Estimating probabilities of multivariate failure sets based on pairwise tail dependence coefficients, arXiv, <doi:10.48550/arXiv.2210.12618>
}
