% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extcoef.R
\name{Xstud.extcoeffun}
\alias{Xstud.extcoeffun}
\title{Extremal coefficient function for the extremal Student process}
\usage{
Xstud.extcoeffun(h, nu, corrfun, ...)
}
\arguments{
\item{h}{vector of pairwise distances}

\item{corrfun}{correlation function}

\item{...}{additional parameters passed to \code{corrfun}}
}
\description{
Extremal coefficient function for the extremal Student process
}
\examples{
extcoef <- Xstud.extcoeffun(seq(0, 10, by = 0.1), nu = 3,
    corrfun = function(x){exp(-x/2)})
plot(extcoef, type = 'l', ylim = c(1,2))
}
