% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{.gev.postpred}
\alias{.gev.postpred}
\title{Posterior predictive distribution and density for the GEV distribution}
\usage{
.gev.postpred(x, posterior, Nyr = 100, type = c("density", "quantile"))
}
\arguments{
\item{x}{\code{n} vector of points}

\item{posterior}{\code{n} by \code{3} matrix of posterior samples}

\item{Nyr}{number of years to extrapolate}

\item{type}{string indicating whether to return the posterior \code{density} or the \code{quantile}.}

\item{vector}{of values for the posterior predictive density or quantile at x}
}
\description{
This function calculates the posterior predictive density at points x
based on a matrix of posterior density parameters
}
\keyword{internal}
