% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmev.R
\name{.is.CNSD}
\alias{.is.CNSD}
\title{Is the matrix conditionally negative semi-definite?
Function adapted from 'is.CNSD' in the CEGO package, v 2.1.0}
\usage{
.is.CNSD(X, tol = 1e-08)
}
\arguments{
\item{X}{a symmetric matrix}

\item{tol}{tolerance value; eigenvalues between \code{-tol} and \code{tol} are assumed to be zero.}
}
\description{
Is the matrix conditionally negative semi-definite?
Function adapted from 'is.CNSD' in the CEGO package, v 2.1.0
}
\author{
Martin Zaefferer
}
\keyword{internal}
