% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.rPexstud}
\alias{.rPexstud}
\title{Generate from extremal Student-t \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function}
\usage{
.rPexstud(index, cholesky, sigma, al)
}
\arguments{
\item{index}{index of the location. An integer in {0, ..., \eqn{d-1}}}

\item{cholesky}{Cholesky root of transformed correlation matrix}

\item{al}{the alpha parameter in Proposition 7. Corresponds to degrees of freedom - 1}

\item{Sigma}{a positive semi-definite correlation matrix}
}
\value{
a \code{d}-vector from \eqn{P_x}
}
\description{
Generate from extremal Student-t \eqn{Y \sim {P_x}}, where
\eqn{P_{x}} is probability of extremal function
}
\keyword{internal}
