% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specdens.R
\name{.wecdf}
\alias{.wecdf}
\title{Weighted empirical distribution function}
\usage{
.wecdf(x, w)
}
\arguments{
\item{x}{observations}

\item{w}{weights}

\item{ord}{logical indicating whether x values are already ordered. Default to \code{FALSE}}
}
\value{
a function of class \code{ecdf}
}
\description{
Compute an empirical distribution function with weights \code{w} at each of \code{x}
}
\keyword{internal}
