% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penultimate.R
\name{egp.fit}
\alias{egp.fit}
\title{Fit of extended GP models and parameter stability plots}
\usage{
egp.fit(xdat, thresh, model = c("egp1", "egp2", "egp3"), init,
  show = FALSE)
}
\description{
This function is an alias of \code{\link{fit.egp}}.
}
\details{
Supported for backward compatibility
}
\keyword{internal}
