% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extgp.R
\name{egp.fit}
\alias{egp.fit}
\title{Fit of extended GP models and parameter stability plots}
\usage{
egp.fit(
  xdat,
  thresh,
  model = c("pt-beta", "pt-gamma", "pt-power", "gj-tnorm", "gj-beta", "exptilt",
    "logist"),
  init,
  show = FALSE
)
}
\description{
This function is an alias of \code{\link{fit.egp}}.
}
\details{
Supported for backward compatibility
}
\keyword{internal}
