% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{gev.bias}
\alias{gev.bias}
\title{Cox-Snell first order bias expression for the GEV distribution}
\usage{
gev.bias(par, n)
}
\arguments{
\item{par}{vector of \code{loc}, \code{scale} and \code{shape}}

\item{n}{sample size}
}
\description{
Bias vector for the GEV distribution based on an \code{n} sample.
}
\seealso{
\code{\link{gev}}
}
\keyword{internal}

