% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R, R/tem.R
\name{gev}
\alias{gev}
\alias{gev.Fscore}
\alias{gev.Vfun}
\alias{gev.bias}
\alias{gev.dphi}
\alias{gev.infomat}
\alias{gev.ll}
\alias{gev.ll.optim}
\alias{gev.phi}
\alias{gev.retlev}
\alias{gev.score}
\title{Generalized extreme value distribution}
\usage{
gev.ll(par, dat)

gev.ll.optim(par, dat)

gev.score(par, dat)

gev.infomat(par, dat, method = c("obs", "exp"))

gev.Vfun(par, dat)

gev.phi(par, dat, V)

gev.dphi(par, dat, V)

gev.bias(par, n)

gev.Fscore(par, dat, method = "obs")

gev.retlev(par, p)
}
\arguments{
\item{par}{vector of parameters, \code{loc}, \code{scale} and \code{shape}}

\item{dat}{sample vector}

\item{method}{string indicating whether to use the expected  (\code{"exp"}) or the observed (\code{"obs"} - the default) information matrix.}

\item{V}{vector calculated by \code{\link{gev.Vfun}}}

\item{n}{sample size}

\item{p}{probability}
}
\description{
Likelihood, score function and information matrix, bias, approximate ancillary statistics and sample space derivative for the generalized extreme value distribution
}
\section{Functions}{
\itemize{
\item \code{gev.ll}: log-likelihood

\item \code{gev.ll.optim}: negative log-likelihood parametrized in terms of \code{loc}, \code{log(scale)} and \code{shape}

\item \code{gev.score}: score vector

\item \code{gev.infomat}: observed or expected information matrix

\item \code{gev.Vfun}: vector implementing conditioning on approximate ancillary statistics for the TEM

\item \code{gev.phi}: canonical parameter in the local exponential family approximation

\item \code{gev.dphi}: derivative matrix of the canonical parameter in the local exponential family approximation

\item \code{gev.bias}: Cox-Snell first order bias

\item \code{gev.Fscore}: Firth's modified score equation

\item \code{gev.retlev}: return level, corresponding to the (1-\eqn{p})th quantile
}}
\note{
The Gumbel case is not currently handled
}
\references{
Cox, D. R. and E. J. Snell (1968). A general definition of residuals, \emph{Journal of the Royal Statistical Society: Series B (Methodological)}, \strong{30}, 248--275.

Cordeiro, G. M. and R. Klein (1994). Bias correction in ARMA models, \emph{Statistics and Probability Letters}, \strong{19}(3), 169--176.

Firth, D. (1993). Bias reduction of maximum likelihood estimates, \emph{Biometrika}, \strong{80}(1), 27--38.
}

