% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpde.infomat}
\alias{gpde.infomat}
\title{Observed information matrix for the GP distribution (expected shortfall)}
\usage{
gpde.infomat(par, dat, m, method = c("obs", "exp"), nobs = length(dat))
}
\arguments{
\item{par}{vector of length 2 containing \eqn{e_m} and \eqn{\xi}, respectively the expected shortfall at probability 1/(1-\eqn{\alpha}) and the shape parameter.}

\item{dat}{sample vector}

\item{m}{number of observations of interest for return levels. See \strong{Details}}

\item{method}{string indicating whether to use the expected  (\code{"exp"}) or the observed (\code{"obs"} - the default) information matrix.}

\item{nobs}{number of observations}
}
\description{
The information matrix is parametrized in terms of excess expected shortfall and shape
}
\seealso{
\code{\link{gpde}}
}
\keyword{internal}
