% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpde.score}
\alias{gpde.score}
\title{Score vector for the GP distribution (expected shortfall)}
\usage{
gpde.score(par, dat, m)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{e_m} and \eqn{\xi}, respectively the expected shortfall at probability 1/(1-\eqn{\alpha}) and the shape parameter.}

\item{dat}{sample vector}

\item{m}{number of observations of interest for return levels. See \strong{Details}}
}
\description{
Score vector for the GP distribution (expected shortfall)
}
\seealso{
\code{\link{gpde}}
}
\keyword{internal}
