% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{gpdr.temstat}
\alias{gpdr.temstat}
\alias{gpdr.Vfun}
\alias{gpdr.phi}
\alias{gpdr.dphi}
\title{Tangent exponential model statistics for the generalized Pareto distribution (return level)}
\usage{
gpdr.Vfun(par, dat, m)

gpdr.phi(par, dat, V, m)

gpdr.dphi(par, dat, V, m)
}
\arguments{
\item{par}{vector of length 2 containing \eqn{y_m} and \eqn{\xi}, respectively the \eqn{m}-year return level and the shape parameter.}

\item{dat}{sample vector}

\item{m}{number of observations of interest for return levels. See \strong{Details}}

\item{V}{vector calculated by \code{gpdr.Vfun}}
}
\description{
Vector implementing conditioning on approximate ancillary statistics for the TEM

Canonical parameter in the local exponential family approximation

Derivative of the canonical parameter \eqn{\phi(\theta)} in the local exponential family approximation
}
\seealso{
\code{\link{gpdr}}

\code{\link{gpdr}}
}
\keyword{internal}
