% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taildep.R
\name{kjtail}
\alias{kjtail}
\title{Estimators of the tail coefficient}
\usage{
kjtail(
  data,
  q,
  ptail = NULL,
  mqu,
  type = 1,
  ties.method = eval(formals(rank)$ties.method)
)
}
\arguments{
\item{data}{a matrix of observations}

\item{q}{vector of quantile levels}

\item{ptail}{tail probability smaller than \code{q}. Default to \code{NULL}}

\item{mqu}{marginal quantile levels for semiparametric estimation; data above this are modelled using a generalized Pareto distribution. If missing, empirical estimation is used throughout}

\item{type}{integer indicating the estimator type}

\item{ties.method}{method for ties}
}
\value{
a list with elements
\itemize{
\item \code{p} quantile level for estimation
\item \code{eta} estimated coefficient of tail dependence \eqn{\eta}
\item \code{eta_sd} estimated standard error of \eqn{\eta}
\item \code{k1} parameter of the tail expansion
\item \code{pat} proportion of observations above the threshold
\item \code{lambda} tail dependence coefficient (sic)
\item \code{tailprob} tail probability, if \code{ptail} is provided
}
}
\description{
Estimators proposed by Krupskii and Joe under second order expansion
for the coefficient of tail dependence \eqn{\eta} and the
joint tail orthant probability
}
\note{
EXPERIMENTAL. The numerical optimization of the likelihood surface is difficult, as the function is ill-behaved. Visual inspection of estimates is necessary to check for non-convergence.
}
\examples{
d <- 2
rho <- 0.9
Sigma <- matrix(rho, d, d) + diag(1 - rho, d)
eta_true <- 1/sum(Sigma)
data <- mev::mvrnorm(
   n = 1e4,
   mu = rep(0, d),
  Sigma = Sigma)
q <- seq(0.95, 0.995, by = 0.005)
taildep <- kjtail(data = data, q = q)
with(taildep,
 plot(x = 1-pat,
      y = eta,
      ylim = c(0,1),
      panel.first = {abline(h = (1+rho)/2)}))
}
