% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taildep2.R
\name{kjtail}
\alias{kjtail}
\title{Estimators of the tail coefficient}
\usage{
kjtail(
  data,
  q,
  ptail = NULL,
  mqu,
  type = 1,
  ties.method = eval(formals(rank)$ties.method),
  ...
)
}
\description{
Estimators proposed by Krupskii and Joe under second order expansion
for the coefficient of tail dependence \eqn{\eta} and the
joint tail orthant probability
}
\examples{
rho <- runif(1, -1, 1)
d <- 2
data <- mev::mvrnorm(
   n = 1e5,
   mu = rep(0, d),
  Sigma = matrix(rho, d, d) + diag(1 - rho, d))
q <- seq(0.95, 0.995, by = 0.005)
taildep <- kjtail(data = data, q = q)
with(taildep,
 plot(x = 1-pat,
      y = eta,
      ylim = c(0,1),
      panel.first = {abline(h = (1+rho)/2)}))
}
