% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univdist.R
\name{rlarg.ll}
\alias{rlarg.ll}
\title{Log-likelihood of the point process of r-largest observations}
\usage{
rlarg.ll(par, dat)
}
\arguments{
\item{par}{vector of \code{loc}, \code{scale} and \code{shape}}

\item{dat}{an \code{n} by \code{r} sample matrix, ordered from largest to smallest in each row}
}
\description{
The function returns the log-likelihood for an \code{n} by \code{r} matrix of observations.
}
\seealso{
\code{\link{rlarg}}
}
\keyword{internal}
