% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{shape.rbm}
\alias{shape.rbm}
\title{Random block maxima shape estimator of Wager}
\usage{
shape.rbm(xdat, k = 10:floor(length(xdat)/2), ...)
}
\arguments{
\item{xdat}{[vector] sample exceedances}

\item{...}{additional parameters, currently ignored}

\item{kmax}{[int] maximum number of observations to consider as sample size}
}
\description{
Computes the shape estimator for varying k up to sample size of maximum \code{kmax} largest observations
}
\note{
Function adapted from **R** package \url{https://github.comswager/rbm}, under MIT license
}
