% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel_exponential_thselect.R
\name{thselect.gbw}
\alias{thselect.gbw}
\title{Kernel-based threshold selection of Goegebeur, Beirlant and de Wet (2008)}
\usage{
thselect.gbw(
  xdat,
  kmax,
  kernel = c("Jackson", "Lewis"),
  rho = c("gbw08", "ghp", "fagh", "cm", "gbw10"),
  ...
)
}
\arguments{
\item{xdat}{vector of observations}

\item{kmax}{maximum number of exceedances considered}

\item{kernel}{[string] kernel choice, one of \code{Jackson} or \code{Lewis}}

\item{rho}{string for the estimator of the second order regular variation. Can also be a negative scalar}
}
\value{

}
\description{
Kernel-based threshold selection of Goegebeur, Beirlant and de Wet (2008)
}
