% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thselect.R
\name{thselect.goks}
\alias{thselect.goks}
\title{Threshold selection based on weighted Kolmogorov-Smirnov distance}
\usage{
thselect.goks(xdat, thresh, B = 100, eps = 0.5)
}
\arguments{
\item{xdat}{vector of observations}

\item{thresh}{vector of thresholds}

\item{B}{number of bootstrap replications}
}
\value{
an object of class \code{mev_thselect_goks} containing
\itemize{
\item{\code{thresh}: ordered vector of candidate thresholds}
\item{\code{thresh0}: selected threshold}
\item{\code{coef}: scale and shape parameters}
\item{\code{nexc}: number of exceedances at each threshold}
\item{\code{stat}: vector of weighted Kolmogorov-Smirnov statistic}
\item{\code{pval}: bootstrap p-value for the weighted Kolmogorov-Smirnov statistic at the selected threshold}
}
}
\description{
Use a semiparametric bootstrap to calculate the null distribution of the weighted Kolmogorov-Smirnov difference between the generalized Pareto distribution and the empirical distribution
}
\examples{
set.seed(2025)
xdat <- rgp(n = 200, shape = 0.1)
thresh <- quantile(xdat, c(0.8,0.9,0.95))
thselect.goks(xdat, thresh, B = 50)
}
\references{
Gonzalo, Jesús and José Olmo (2004). Which Extreme Values Are Really Extreme?, \emph{Journal of Financial Econometrics}, 2(\bold{3}), <doi:10.1093/jjfinec/nbh014>
}
