% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailindex.R
\name{thselect.rbm}
\alias{thselect.rbm}
\title{Threshold selection for the random block maxima method}
\usage{
thselect.rbm(xdat, kmax = length(xdat))
}
\arguments{
\item{xdat}{[vector] sample exceedances}

\item{kmax}{maximum number of exceedances to consider.}
}
\value{
a list with elements
}
\description{
Threshold selection for the random block maxima method
}
