% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thselect.R
\name{thselect.wcvm}
\alias{thselect.wcvm}
\title{Threshold selection via minimization of the weighted Cramer-von Mises distance}
\usage{
thselect.wcvm(xdat, k)
}
\arguments{
\item{xdat}{vector of positive exceedances}

\item{k}{vector of number of exceedances, or integer indicating the maximum value of \eqn{k}, in which case a vector of integers from \eqn{k=10} to \code{k} is constructed}
}
\value{
an object of class \code{mev_thselect_wcvm} (list) with elements
\itemize{
\item \code{k0}: selected number of order statistics
\item \code{shape}: Hill estimate of the shape at selected threshold
\item \code{thresh}: value of the threshold (the (k+1)st largest order statistic)
\item \code{criterion}: a data frame with columns \code{k} and \code{crit} giving the criterion value
}
}
\description{
For a Pareto-type sample, return the threshold that
minimizes a weighted Cramer-von Mises criterion for the
exponential sample with scale \eqn{H_{n, n_u}} and
the log increments.
}
\references{
Goegebeur , Y., Beirlant , J., and de Wet , T. (2008). Linking Pareto-Tail Kernel Goodness-of-fit Statistics with Tail Index at Optimal Threshold and Second Order Estimation. REVSTAT-Statistical Journal, 6(\bold{1}), 51-69. <doi:10.57805/revstat.v6i1.57>
}
